test: de-duplicate and shard @rjsf/core's slowest test suites
- Split the ~3.7k-line "Form common" suite out of Form.test.tsx into six
  topical test files (rendering, defaults, handlers, state, errors,
  props), each still run under all four omitExtraData/liveOmit variants
  via the existing describeRepeated helper. vitest can now parallelize
  what used to run serially in one file.
- Stop re-running the whole StringField suite inside ArrayField/ObjectField
  via test-file cross-imports; the shared cases live in TextWidgetTest.tsx
  and run once. Removes 302 duplicate test executions.

@rjsf/core: 2240 tests / ~10.2s -> 1938 tests / ~5.4s locally, identical
coverage.
