+++ CHANGELOG.md
@@ -128,6 +128,7 @@ should change the heading of the (upcoming) version to include a major version b
 - Removed the last `lodash` usage from the playground and the test suites, so no package declares `lodash`/`lodash-es` as a direct dependency any more (both remain in the lockfile as transitive dependencies of third-party tooling). The `lodashReplacer` `tsc-alias` machinery that rewrote `lodash` imports to `lodash-es` for the ESM builds has been deleted along with it
 - Cleaned up `@rjsf/validator-ajv8`'s tests: the deliberately-triggered AJV `console.warn` output is now asserted with a shared `expectWarn()` helper instead of leaking into the test output, the gitignored `superSchema*.cjs` harness files are regenerated by a vitest `globalSetup` (so watch mode and IDE runners work from a fresh clone too), and the harness files are read relative to the test file instead of the working directory
 - Added `size-limit` CI checks that enforce bundle size budgets for `@rjsf/core`, `@rjsf/utils` and `@rjsf/validator-ajv8` — each measured both with and without its runtime dependencies, plus single-export tree-shaking canaries — and comment the size diff versus the base branch on every PR
+- Restructured `@rjsf/core`'s test suite for speed: the "Form common" suite was split by topic into six test files so vitest can parallelize them, and the `StringField` suite is no longer re-executed via cross-imports from the `ArrayField`/`ObjectField` suites (302 duplicate test executions removed)
 
 # 6.8.0
