+++ packages/core/test/Form.test.tsx
@@ -1,4002 +1,27 @@
 import { createRef, useEffect, useRef, useState, useCallback } from 'react';
-import type {
-  ErrorSchema,
-  Experimental_DefaultFormStateBehavior,
-  FieldProps,
-  FieldTemplateProps,
-  FormValidation,
-  RJSFSchema,
-  UiSchema,
-  ValidatorType,
-  WidgetProps,
-} from '@rjsf/utils';
-import {
-  bracketNameGenerator,
-  buttonId,
-  dotNotationNameGenerator,
-  getTemplate,
-  getUiOptions,
-  optionalControlsId,
-  noop,
-} from '@rjsf/utils';
-import validator, { customizeValidator } from '@rjsf/validator-ajv8';
-import { act, render, waitFor } from '@testing-library/react';
+import type { ErrorSchema, Experimental_DefaultFormStateBehavior, FieldProps, RJSFSchema, UiSchema } from '@rjsf/utils';
+import { bracketNameGenerator, buttonId, dotNotationNameGenerator, optionalControlsId } from '@rjsf/utils';
+import validator from '@rjsf/validator-ajv8';
+import { act, render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import draft06 from 'ajv/lib/refs/json-schema-draft-06.json';
-import { createPortal } from 'react-dom';
-import type { Mock } from 'vitest';
 
 import type { FormProps, IChangeEvent } from '../src';
 import Form from '../src';
-import type { NoValFormProps, RerenderType } from './testUtils';
+import type { NoValFormProps } from './testUtils';
 import {
   actWrappedDelayPromise,
   createComponent,
   createFormComponent,
   delayPromise,
-  describeRepeated,
   expectToHaveBeenCalledWithFormData,
-  renderNode,
   setupConsoleErrorSuppression,
   submitForm,
 } from './testUtils';
 import widgetsSchema from './widgets_schema.json';
 
-const TWO_BUTTONS = (
-  <>
-    <button type='submit'>Submit</button>
-    <button type='submit'>Another submit</button>
-  </>
-);
 const user = userEvent.setup();
 
-const renderErrorSuppression = setupConsoleErrorSuppression();
-
-describeRepeated('Form common', (createFormComponent) => {
-  describe('Empty schema', () => {
-    it('Should throw error when Form is missing validator', () => {
-      expect(() =>
-        createComponent(Form, { ref: createRef(), schema: {}, validator: undefined as unknown as ValidatorType }),
-      ).toThrow('A validator is required for Form functionality to work');
-      expect(renderErrorSuppression.consoleSpy).toHaveBeenCalledWith(
-        expect.stringContaining('The above error occurred in the <Form> component'),
-      );
-    });
-
-    it('should render a form tag', () => {
-      const { node } = createFormComponent({ ref: createRef(), schema: {} });
-
-      expect(node.tagName).toEqual('FORM');
-    });
-
-    it('should render a submit button', () => {
-      const { node } = createFormComponent({ ref: createRef(), schema: {} });
-
-      expect(node.querySelectorAll('button[type=submit]')).toHaveLength(1);
-    });
-
-    it('should render children buttons', () => {
-      const { node } = createFormComponent({
-        ref: createRef(),
-        schema: {},
-        children: TWO_BUTTONS,
-      });
-      expect(node.querySelectorAll('button[type=submit]')).toHaveLength(2);
-    });
-
-    it("should render errors if schema isn't object", () => {
-      const { node } = createFormComponent({
-        ref: createRef(),
-        schema: {
-          type: 'object',
-          title: 'object',
-          properties: {
-            firstName: 'some mame',
-            address: {
-              $ref: '#/definitions/address',
-            },
-          },
-          definitions: {
-            address: {
-              street: 'some street',
-            },
-          },
-        } as RJSFSchema,
-      });
-      expect(node.querySelector('.unsupported-field')).toHaveTextContent('Unknown field type undefined');
-    });
-
-    it('will render fallback ui when useFallbackUiForUnsupportedType is true', async () => {
-      const schema = {
-        type: 'object',
-        title: 'object',
-        properties: {
-          unknownProperty: {
-            type: 'someUnsupportedType',
-          },
-        },
-      } as unknown as RJSFSchema;
-      const props: NoValFormProps = {
-        useFallbackUiForUnsupportedType: true,
-        schema,
-        formData: {
-          unknownProperty: '123456',
-        },
-      };
-
-      const { node, onChange } = createFormComponent({ ...props });
-
-      expect(node.querySelectorAll('.unsupported-field')).toHaveLength(0);
-      expect(node.querySelector('select')).toBeInTheDocument();
-      let select = node.querySelector('select')!;
-      let options = node.querySelectorAll<HTMLOptionElement>('select option');
-      expect(options).toHaveLength(5);
-      expect(options[0]).toHaveTextContent('string');
-      expect(options[0].selected).toBe(true);
-      expect(node.querySelector<HTMLInputElement>('input[type=text]')!).toHaveAttribute('value', '123456');
-
-      // Change the fallback type to 'number'
-      await user.selectOptions(select, options[1]);
-      expect(options[1]).toHaveTextContent('number');
-      expect(options[1].selected).toBe(true);
-      expect(node.querySelector<HTMLInputElement>('input[type=number]')).toBeInTheDocument();
-      expect(node.querySelector<HTMLInputElement>('input[type=number]')).toHaveAttribute('value', '123456');
-
-      // Verify formData was casted to number
-      expectToHaveBeenCalledWithFormData(onChange, { unknownProperty: 123456 }, 'root_unknownProperty');
-
-      select = node.querySelector('select')!;
-      options = node.querySelectorAll<HTMLOptionElement>('select option');
-      // Change the fallback type to 'boolean'
-      await user.selectOptions(select, options[2]);
-      expect(options[2]).toHaveTextContent('boolean');
-      expect(options[2].selected).toBe(true);
-      expect(node.querySelector<HTMLInputElement>('input[type=checkbox]')).toBeInTheDocument();
-      expect(node.querySelector<HTMLInputElement>('input[type=checkbox]')).toBeChecked();
-      // Verify formData was casted to number
-      expectToHaveBeenCalledWithFormData(onChange, { unknownProperty: true }, 'root_unknownProperty');
-
-      select = node.querySelector('select')!;
-      options = node.querySelectorAll<HTMLOptionElement>('select option');
-      // Change the fallback type to 'object'
-      await user.selectOptions(select, options[3]);
-      expect(options[3]).toHaveTextContent('object');
-      expect(options[3].selected).toBe(true);
-      let addButton = node.querySelector<HTMLButtonElement>('.rjsf-object-property-expand button');
-      expect(addButton).toBeInTheDocument();
-      // click the add button
-      await user.click(addButton!);
-
-      // Verify formData was casted to object
-      expectToHaveBeenCalledWithFormData(
-        onChange,
-        { unknownProperty: { newKey: 'New Value' } },
-        'root_unknownProperty',
-      );
-
-      select = node.querySelector('select')!;
-      options = node.querySelectorAll<HTMLOptionElement>('select option');
-      // Change the fallback type to 'array'
-      await user.selectOptions(select, options[4]);
-      expect(options[4]).toHaveTextContent('array');
-      expect(options[4].selected).toBe(true);
-      addButton = node.querySelector<HTMLButtonElement>('.rjsf-array-item-add button');
-      expect(addButton).toBeInTheDocument();
-      // click the add button
-      await user.click(addButton!);
-
-      // Verify formData was casted to array
-      expectToHaveBeenCalledWithFormData(onChange, { unknownProperty: [undefined] }, 'root_unknownProperty');
-    });
-  });
-
-  describe('on component creation', () => {
-    const schema: RJSFSchema = {
-      type: 'object',
-      title: 'root object',
-      required: ['count'],
-      properties: {
-        count: {
-          type: 'number',
-          default: 789,
-        },
-      },
-    };
-
-    describe('when props.formData does not equal the default values', () => {
-      it('should call props.onChange with current state', () => {
-        const formData = {
-          foo: 123,
-        };
-        const { onChange } = createFormComponent({ schema, formData });
-        expect(onChange).toHaveBeenCalledTimes(1);
-        expectToHaveBeenCalledWithFormData(onChange, { ...formData, count: 789 });
-      });
-    });
-
-    describe('when props.formData equals the default values', () => {
-      it('should not call props.onChange', () => {
-        const formData = {
-          count: 789,
-        };
-        const { onChange } = createFormComponent({ schema, formData });
-        expect(onChange).not.toHaveBeenCalled();
-      });
-    });
-  });
-
-  describe('Option idPrefix', () => {
-    it('should change the rendered ids', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        title: 'root object',
-        required: ['foo'],
-        properties: {
-          count: {
-            type: 'number',
-          },
-        },
-      };
-      const { node } = createFormComponent({ schema, idPrefix: 'rjsf' });
-      const inputs = node.querySelectorAll('input');
-      const ids = [];
-      for (let i = 0, len = inputs.length; i < len; i += 1) {
-        const input = inputs[i];
-        ids.push(input.getAttribute('id'));
-      }
-      expect(ids).toEqual(['rjsf_count']);
-      expect(node.querySelector('fieldset')).toHaveAttribute('id', 'rjsf');
-    });
-  });
-
-  describe('Changing idPrefix', () => {
-    it('should work with simple example', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        title: 'root object',
-        required: ['foo'],
-        properties: {
-          count: {
-            type: 'number',
-          },
-        },
-      };
-      const { node } = createFormComponent({ schema, idPrefix: 'rjsf' });
-      const inputs = node.querySelectorAll('input');
-      const ids = [];
-      for (let i = 0, len = inputs.length; i < len; i += 1) {
-        const input = inputs[i];
-        ids.push(input.getAttribute('id'));
-      }
-      expect(ids).toEqual(['rjsf_count']);
-      expect(node.querySelector('fieldset')).toHaveAttribute('id', 'rjsf');
-    });
-
-    it('should work with oneOf', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          connector: {
-            type: 'string',
-            enum: ['aws', 'gcp'],
-            title: 'Provider',
-            default: 'aws',
-          },
-        },
-        dependencies: {
-          connector: {
-            oneOf: [
-              {
-                type: 'object',
-                properties: {
-                  connector: {
-                    type: 'string',
-                    enum: ['aws'],
-                  },
-                  key_aws: {
-                    type: 'string',
-                  },
-                },
-              },
-              {
-                type: 'object',
-                properties: {
-                  connector: {
-                    type: 'string',
-                    enum: ['gcp'],
-                  },
-                  key_gcp: {
-                    type: 'string',
-                  },
-                },
-              },
-            ],
-          },
-        },
-      };
-
-      const { node } = createFormComponent({ schema, idPrefix: 'rjsf' });
-      const inputs = node.querySelectorAll('input');
-      const ids = [];
-      for (let i = 0, len = inputs.length; i < len; i += 1) {
-        const input = inputs[i];
-        ids.push(input.getAttribute('id'));
-      }
-      expect(ids).toEqual(['rjsf_key_aws']);
-    });
-  });
-
-  describe('Option idSeparator', () => {
-    it('should change the rendered ids', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        title: 'root object',
-        required: ['foo'],
-        properties: {
-          count: {
-            type: 'number',
-          },
-        },
-      };
-      const { node } = createFormComponent({ schema, idSeparator: '.' });
-      const inputs = node.querySelectorAll('input');
-      const ids = [];
-      for (let i = 0, len = inputs.length; i < len; i += 1) {
-        const input = inputs[i];
-        ids.push(input.getAttribute('id'));
-      }
-      expect(ids).toEqual(['root.count']);
-    });
-  });
-
-  describe('Custom field template', () => {
-    const schema: RJSFSchema = {
-      type: 'object',
-      title: 'root object',
-      required: ['foo'],
-      properties: {
-        foo: {
-          type: 'string',
-          description: 'this is description',
-          minLength: 32,
-        },
-      },
-    };
-
-    const uiSchema: UiSchema = {
-      foo: {
-        'ui:help': 'this is help',
-      },
-    };
-
-    const formData = { foo: 'invalid' };
-
-    function CustomFieldTemplate(props: FieldTemplateProps) {
-      const {
-        id,
-        classNames,
-        label,
-        help,
-        rawHelp,
-        required,
-        description,
-        rawDescription,
-        errors,
-        rawErrors,
-        children,
-      } = props;
-      return (
-        <div className={`my-template ${classNames}`}>
-          <label htmlFor={id}>
-            {label}
-            {required ? '*' : null}
-          </label>
-          {description}
-          {children}
-          {errors}
-          {help}
-          <span className='raw-help'>{`${rawHelp} rendered from the raw format`}</span>
-          <span className='raw-description'>{`${rawDescription} rendered from the raw format`}</span>
-          {rawErrors ? (
-            <ul>
-              {rawErrors.map((error, i) => (
-                // oxlint-disable-next-line react/no-array-index-key
-                <li key={i} className='raw-error'>
-                  {error}
-                </li>
-              ))}
-            </ul>
-          ) : null}
-        </div>
-      );
-    }
-
-    let node: Element;
-
-    beforeEach(() => {
-      node = createFormComponent({
-        schema,
-        uiSchema,
-        formData,
-        templates: {
-          FieldTemplate: CustomFieldTemplate,
-        },
-        liveValidate: true,
-      }).node;
-    });
-
-    it('should use the provided field template', () => {
-      expect(node.querySelector('.my-template')).toBeInTheDocument();
-    });
-
-    it('should use the provided template for labels', () => {
-      expect(node.querySelector('.my-template > label')).toHaveTextContent('root object');
-      expect(node.querySelector('.my-template .rjsf-field-string > label')).toHaveTextContent('foo*');
-    });
-
-    it('should pass description as the provided React element', () => {
-      expect(node.querySelector('#root_foo__description')).toHaveTextContent('this is description');
-    });
-
-    it('should pass rawDescription as a string', () => {
-      expect(node.querySelector('.raw-description')).toHaveTextContent(
-        'this is description rendered from the raw format',
-      );
-    });
-
-    it('should pass errors as the provided React component', async () => {
-      // live validate does not run on initial render anymore
-      expect(node.querySelectorAll('.error-detail li')).toHaveLength(0);
-      const input = node.querySelector<HTMLInputElement>('input')!;
-      await user.clear(input);
-      await user.type(input, 'stillinvalid');
-      expect(node.querySelectorAll('.error-detail li')).toHaveLength(1);
-    });
-
-    it('should pass rawErrors as an array of strings', async () => {
-      // live validate does not run on initial render anymore
-      expect(node.querySelectorAll('.raw-error')).toHaveLength(0);
-      const input = node.querySelector<HTMLInputElement>('input')!;
-      await user.clear(input);
-      await user.type(input, 'stillinvalid');
-      expect(node.querySelectorAll('.raw-error')).toHaveLength(1);
-    });
-
-    it('should pass help as a the provided React element', () => {
-      expect(node.querySelector('.help-block')).toHaveTextContent('this is help');
-    });
-
-    it('should pass rawHelp as a string', () => {
-      expect(node.querySelector('.raw-help')).toHaveTextContent('this is help rendered from the raw format');
-    });
-  });
-
-  describe('ui options submitButtonOptions', () => {
-    it('should not render a submit button', () => {
-      const props: NoValFormProps = {
-        schema: {},
-        uiSchema: { 'ui:submitButtonOptions': { norender: true } },
-      };
-      const { node } = createFormComponent(props);
-      expect(node.querySelectorAll('button[type=submit]')).toHaveLength(0);
-    });
-
-    it('should render a submit button with text Confirm', () => {
-      const props: NoValFormProps = {
-        schema: {},
-        uiSchema: { 'ui:submitButtonOptions': { submitText: 'Confirm' } },
-      };
-      const { node } = createFormComponent(props);
-      expect(node.querySelector('button[type=submit]')).toHaveTextContent('Confirm');
-    });
-  });
-
-  describe('Custom submit buttons', () => {
-    // Submit events on buttons are not fired on disconnected forms
-    // So we need to add the DOM tree to the body in this case.
-    // See: https://github.com/jsdom/jsdom/pull/1865
-    // https://developer.mozilla.org/en-US/docs/Web/API/Node/isConnected
-    const domNode = document.createElement('div');
-    beforeEach(() => {
-      document.body.appendChild(domNode);
-    });
-    afterEach(() => {
-      document.body.removeChild(domNode);
-    });
-    it('should submit the form when clicked', async () => {
-      const { node, onSubmit } = createFormComponent({ schema: {}, children: TWO_BUTTONS });
-      const buttons = node.querySelectorAll<HTMLButtonElement>('button[type=submit]');
-      expect(buttons).toHaveLength(2);
-      await user.click(buttons[0]);
-      await user.click(buttons[1]);
-      expect(onSubmit).toHaveBeenCalledTimes(2);
-    });
-  });
-
-  describe('Schema definitions', () => {
-    it('should use a single schema definition reference', () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          testdef: { type: 'string' },
-        },
-        $ref: '#/definitions/testdef',
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelectorAll('input[type=text]')).toHaveLength(1);
-    });
-
-    it('should handle multiple schema definition references', () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          testdef: { type: 'string' },
-        },
-        type: 'object',
-        properties: {
-          foo: { $ref: '#/definitions/testdef' },
-          bar: { $ref: '#/definitions/testdef' },
-        },
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelectorAll('input[type=text]')).toHaveLength(2);
-    });
-
-    it('should handle deeply referenced schema definitions', () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          testdef: { type: 'string' },
-        },
-        type: 'object',
-        properties: {
-          foo: {
-            type: 'object',
-            properties: {
-              bar: { $ref: '#/definitions/testdef' },
-            },
-          },
-        },
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelectorAll('input[type=text]')).toHaveLength(1);
-    });
-
-    it('should handle references to deep schema definitions', () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          testdef: {
-            type: 'object',
-            properties: {
-              bar: { type: 'string' },
-            },
-          },
-        },
-        type: 'object',
-        properties: {
-          foo: { $ref: '#/definitions/testdef/properties/bar' },
-        },
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelectorAll('input[type=text]')).toHaveLength(1);
-    });
-
-    it('should handle referenced definitions for array items', () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          testdef: { type: 'string' },
-        },
-        type: 'object',
-        properties: {
-          foo: {
-            type: 'array',
-            items: { $ref: '#/definitions/testdef' },
-          },
-        },
-      };
-
-      const { node } = createFormComponent({
-        schema,
-        formData: {
-          foo: ['blah'],
-        },
-      });
-
-      expect(node.querySelectorAll('input[type=text]')).toHaveLength(1);
-    });
-
-    it('should not crash with null values for property with additionalProperties', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          data: {
-            additionalProperties: {
-              type: 'string',
-            },
-            type: 'object',
-          },
-        },
-      };
-
-      const { node } = createFormComponent({
-        schema,
-        formData: {
-          data: null,
-        },
-      });
-
-      expect(node).not.toBeNull();
-    });
-
-    it('should not crash with non-object values for property with additionalProperties', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          data1: {
-            additionalProperties: {
-              type: 'string',
-            },
-            type: 'object',
-          },
-          data2: {
-            additionalProperties: {
-              type: 'string',
-            },
-            type: 'object',
-          },
-        },
-      };
-
-      const { node } = createFormComponent({
-        schema,
-        formData: {
-          data1: 123,
-          data2: ['one', 'two', 'three'],
-        },
-      });
-
-      expect(node).not.toBeNull();
-    });
-
-    it('should raise for non-existent definitions referenced', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          foo: { $ref: '#/definitions/nonexistent' },
-        },
-      };
-
-      expect(() => createFormComponent({ schema })).toThrow(/#\/definitions\/nonexistent/);
-      expect(renderErrorSuppression.consoleSpy).toHaveBeenCalledWith(
-        expect.stringContaining('The above error occurred in the <Form> component'),
-      );
-    });
-
-    it('should propagate referenced definition defaults', () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          testdef: { type: 'string', default: 'hello' },
-        },
-        $ref: '#/definitions/testdef',
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelector<HTMLInputElement>('input[type=text]')!).toHaveValue('hello');
-    });
-
-    it('should propagate nested referenced definition defaults', () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          testdef: { type: 'string', default: 'hello' },
-        },
-        type: 'object',
-        properties: {
-          foo: { $ref: '#/definitions/testdef' },
-        },
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelector<HTMLInputElement>('input[type=text]')!).toHaveValue('hello');
-    });
-
-    it('should propagate referenced definition defaults for array items', async () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          testdef: { type: 'string', default: 'hello' },
-        },
-        type: 'array',
-        items: {
-          $ref: '#/definitions/testdef',
-        },
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      await user.click(node.querySelector('.rjsf-array-item-add button')!);
-
-      expect(node.querySelector<HTMLInputElement>('input[type=text]')!).toHaveValue('hello');
-    });
-
-    it('should propagate referenced definition defaults in objects with additionalProperties', async () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          testdef: { type: 'string' },
-        },
-        type: 'object',
-        additionalProperties: {
-          $ref: '#/definitions/testdef',
-        },
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      await user.click(node.querySelector('.btn-add')!);
-
-      expect(node.querySelector<HTMLInputElement>('input[type=text]')!).toHaveValue('newKey');
-    });
-
-    it('should propagate referenced definition defaults in objects with additionalProperties that have a type present', async () => {
-      // Though `additionalProperties` has a `type` present here, it also has a `$ref` so that
-      // referenced schema should override it.
-      const schema: RJSFSchema = {
-        definitions: {
-          testdef: { type: 'number' },
-        },
-        type: 'object',
-        additionalProperties: {
-          type: 'string',
-          $ref: '#/definitions/testdef',
-        },
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      await user.click(node.querySelector('.btn-add')!);
-
-      expect(node.querySelector<HTMLInputElement>('input[type=number]')).toHaveValue(0);
-    });
-
-    it('should recursively handle referenced definitions', async () => {
-      const schema: RJSFSchema = {
-        $ref: '#/definitions/node',
-        definitions: {
-          node: {
-            type: 'object',
-            properties: {
-              name: { type: 'string' },
-              children: {
-                type: 'array',
-                items: {
-                  $ref: '#/definitions/node',
-                },
-              },
-            },
-          },
-        },
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelector('#root_children_0_name')).not.toBeInTheDocument();
-
-      await user.click(node.querySelector('.rjsf-array-item-add button')!);
-
-      expect(node.querySelector('#root_children_0_name')).toBeInTheDocument();
-    });
-
-    it('should follow recursive references', () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          bar: { $ref: '#/definitions/qux' },
-          qux: { type: 'string' },
-        },
-        type: 'object',
-        required: ['foo'],
-        properties: {
-          foo: { $ref: '#/definitions/bar' },
-        },
-      };
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelectorAll('input[type=text]')).toHaveLength(1);
-    });
-
-    it('should follow multiple recursive references', () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          bar: { $ref: '#/definitions/bar2' },
-          bar2: { $ref: '#/definitions/qux' },
-          qux: { type: 'string' },
-        },
-        type: 'object',
-        required: ['foo'],
-        properties: {
-          foo: { $ref: '#/definitions/bar' },
-        },
-      };
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelectorAll('input[type=text]')).toHaveLength(1);
-    });
-
-    it('should priorize definition over schema type property', () => {
-      // Refs bug #140
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          name: { type: 'string' },
-          childObj: {
-            type: 'object',
-            $ref: '#/definitions/childObj',
-          },
-        },
-        definitions: {
-          childObj: {
-            type: 'object',
-            properties: {
-              otherName: { type: 'string' },
-            },
-          },
-        },
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelectorAll('input[type=text]')).toHaveLength(2);
-    });
-
-    it('should priorize local properties over definition ones', () => {
-      // Refs bug #140
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          foo: {
-            title: 'custom title',
-            $ref: '#/definitions/objectDef',
-          },
-        },
-        definitions: {
-          objectDef: {
-            type: 'object',
-            title: 'definition title',
-            properties: {
-              field: { type: 'string' },
-            },
-          },
-        },
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelector('legend')).toHaveTextContent('custom title');
-    });
-
-    it('should propagate and handle a resolved schema definition', () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          enumDef: { type: 'string', enum: ['a', 'b'] },
-        },
-        type: 'object',
-        properties: {
-          name: { $ref: '#/definitions/enumDef' },
-        },
-      };
-
-      const { node } = createFormComponent({ schema });
-
-      expect(node.querySelectorAll('option')).toHaveLength(3);
-    });
-  });
-
-  describe('Default value handling on clear', () => {
-    const schema: RJSFSchema = {
-      type: 'string',
-      default: 'foo',
-    };
-
-    it('should not set default when a text field is cleared', async () => {
-      const { node } = createFormComponent({ schema, formData: 'bar' });
-
-      await user.clear(node.querySelector<HTMLInputElement>('input')!);
-
-      expect(node.querySelector<HTMLInputElement>('input')).toHaveValue('');
-    });
-  });
-
-  describe('Defaults array items default propagation', () => {
-    const schema: RJSFSchema = {
-      type: 'object',
-      title: 'lvl 1 obj',
-      properties: {
-        object: {
-          type: 'object',
-          title: 'lvl 2 obj',
-          properties: {
-            array: {
-              type: 'array',
-              items: {
-                type: 'object',
-                title: 'lvl 3 obj',
-                properties: {
-                  bool: {
-                    type: 'boolean',
-                    default: true,
-                  },
-                },
-              },
-            },
-          },
-        },
-      },
-    };
-
-    it('should propagate deeply nested defaults to submit handler', async () => {
-      const { node, onSubmit } = createFormComponent({ schema });
-
-      await user.click(node.querySelector('.rjsf-array-item-add button')!);
-      await submitForm(node, user);
-
-      expectToHaveBeenCalledWithFormData(onSubmit, { object: { array: [{ bool: true }] } }, true);
-    });
-  });
-
-  describe('Defaults additionalProperties propagation', () => {
-    it('should submit string string map defaults', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        additionalProperties: {
-          type: 'string',
-        },
-        default: {
-          foo: 'bar',
-        },
-      };
-
-      const { node, onSubmit } = createFormComponent({ schema });
-      await submitForm(node, user);
-
-      expectToHaveBeenCalledWithFormData(onSubmit, { foo: 'bar' }, true);
-    });
-
-    it('should submit a combination of properties and additional properties defaults', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          x: {
-            type: 'string',
-          },
-        },
-        additionalProperties: {
-          type: 'string',
-        },
-        default: {
-          x: 'x default value',
-          y: 'y default value',
-        },
-      };
-
-      const { node, onSubmit } = createFormComponent({ schema });
-      await submitForm(node, user);
-
-      expectToHaveBeenCalledWithFormData(onSubmit, { x: 'x default value', y: 'y default value' }, true);
-    });
-
-    it('should submit a properties and additional properties defaults when properties default is nested', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          x: {
-            type: 'string',
-            default: 'x default value',
-          },
-        },
-        additionalProperties: {
-          type: 'string',
-        },
-        default: {
-          y: 'y default value',
-        },
-      };
-
-      const { node, onSubmit } = createFormComponent({ schema });
-      await submitForm(node, user);
-
-      expectToHaveBeenCalledWithFormData(onSubmit, { x: 'x default value', y: 'y default value' }, true);
-    });
-
-    it('should submit defaults when nested map has map values', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          x: {
-            additionalProperties: {
-              $ref: '#/definitions/objectDef',
-            },
-          },
-        },
-        definitions: {
-          objectDef: {
-            type: 'object',
-            additionalProperties: {
-              type: 'string',
-            },
-          },
-        },
-        default: {
-          x: {
-            y: {
-              z: 'x.y.z default value',
-            },
-          },
-        },
-      };
-
-      const { node, onSubmit } = createFormComponent({ schema });
-      await submitForm(node, user);
-
-      expectToHaveBeenCalledWithFormData(onSubmit, { x: { y: { z: 'x.y.z default value' } } }, true);
-    });
-
-    it('should submit defaults when they are defined in a nested additionalProperties', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          x: {
-            additionalProperties: {
-              type: 'string',
-              default: 'x.y default value',
-            },
-          },
-        },
-        default: {
-          x: {
-            y: {},
-          },
-        },
-      };
-
-      const { node, onSubmit } = createFormComponent({ schema });
-      await submitForm(node, user);
-
-      expectToHaveBeenCalledWithFormData(onSubmit, { x: { y: 'x.y default value' } }, true);
-    });
-
-    it('should submit defaults when additionalProperties is a boolean value', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        additionalProperties: true,
-        default: {
-          foo: 'bar',
-        },
-      };
-
-      const { node, onSubmit } = createFormComponent({ schema });
-      await submitForm(node, user);
-
-      expectToHaveBeenCalledWithFormData(onSubmit, { foo: 'bar' }, true);
-    });
-
-    it('should NOT submit default values when additionalProperties is false', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          foo: {
-            type: 'string',
-          },
-        },
-        additionalProperties: false,
-        default: {
-          foo: "I'm the only one",
-          bar: "I don't belong here",
-        },
-      };
-
-      const { node, onSubmit } = createFormComponent({ schema });
-      await submitForm(node, user);
-
-      expectToHaveBeenCalledWithFormData(onSubmit, { foo: "I'm the only one" }, true);
-    });
-  });
-
-  describe('Submit handler', () => {
-    it('should call provided submit handler with form state', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          foo: { type: 'string' },
-        },
-      };
-      const formData = {
-        foo: 'bar',
-      };
-      const { node, onSubmit } = createFormComponent({
-        ref: createRef(),
-        schema,
-        formData,
-      });
-
-      await submitForm(node, user);
-      expectToHaveBeenCalledWithFormData(onSubmit, formData, true);
-    });
-
-    it('should not call provided submit handler on validation errors', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          foo: {
-            type: 'string',
-            minLength: 1,
-          },
-        },
-      };
-      const formData = {
-        foo: '',
-      };
-      const { node, onSubmit, onError } = createFormComponent({
-        ref: createRef(),
-        schema,
-        formData,
-      });
-
-      await submitForm(node, user);
-
-      expect(onSubmit).not.toHaveBeenCalled();
-      expect(onError).toHaveBeenCalled();
-    });
-  });
-
-  describe('Change handler', () => {
-    it('should call provided change handler on form state change with schema and uiSchema', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          foo: {
-            type: 'string',
-          },
-        },
-      };
-      const uiSchema: UiSchema = {
-        foo: { 'ui:field': 'textarea' },
-      };
-
-      const formData = {
-        foo: '',
-      };
-      const { node, onChange } = createFormComponent({
-        ref: createRef(),
-        schema,
-        uiSchema,
-        formData,
-      });
-
-      await user.type(node.querySelector('[type=text]')!, 'new');
-
-      expectToHaveBeenCalledWithFormData(onChange, { foo: 'new' }, 'root_foo');
-    });
-    it('should call last provided change handler', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          foo: {
-            type: 'string',
-            default: 'bar',
-          },
-        },
-      };
-
-      const secondOnChange = vi.fn();
-
-      const { onChange, rerender } = createFormComponent({ ref: createRef(), schema, formData: { foo: 'bar1' } });
-
-      rerender({ schema, formData: {}, onChange });
-
-      expect(onChange).toHaveBeenCalledTimes(1);
-
-      rerender({ schema, formData: { foo: 'bar2' } });
-
-      rerender({ schema, formData: {}, onChange: secondOnChange });
-
-      expect(onChange).toHaveBeenCalledTimes(1);
-      expect(secondOnChange).toHaveBeenCalledTimes(1);
-    });
-    it('should call change handler with proper data after two near simultaneous changes', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          foo: {
-            type: 'string',
-            default: 'bar',
-          },
-          baz: {
-            type: 'string',
-            default: 'blah',
-          },
-        },
-      };
-      function FooWidget(props: WidgetProps) {
-        const { value, id, onChange, uiSchema, registry } = props;
-        const uiOptions = getUiOptions(uiSchema);
-        const BaseInputTemplate = getTemplate('BaseInputTemplate', registry, uiOptions);
-        useEffect(() => {
-          if (value === 'bar') {
-            onChange('bar2', undefined, id);
-          }
-        }, [value, onChange, id]);
-        return <BaseInputTemplate {...props} />;
-      }
-      function BazWidget(props: WidgetProps) {
-        const { value, id, onChange, uiSchema, registry } = props;
-        const uiOptions = getUiOptions(uiSchema);
-        const BaseInputTemplate = getTemplate('BaseInputTemplate', registry, uiOptions);
-        useEffect(() => {
-          if (value === 'blah') {
-            onChange('blah2', undefined, id);
-          }
-        }, [value, onChange, id]);
-        return <BaseInputTemplate {...props} />;
-      }
-      const uiSchema: UiSchema = {
-        foo: {
-          'ui:widget': FooWidget,
-        },
-        baz: {
-          'ui:widget': BazWidget,
-        },
-      };
-
-      let formData = {};
-      const ids: (string | undefined)[] = [];
-      const onChange: FormProps['onChange'] = (data, id) => {
-        const { formData: fd } = data;
-        formData = { ...formData, ...fd };
-        ids.push(id);
-      };
-      createFormComponent({
-        schema,
-        formData,
-        onChange,
-        uiSchema,
-      });
-
-      await waitFor(() => {
-        expect(ids).toHaveLength(3);
-      });
-
-      expect(formData).toEqual({ foo: 'bar2', baz: 'blah2' });
-      // There will be 3 ids, undefined for the setting of the defaults and then the two updated components
-      expect(ids).toEqual([undefined, 'root_foo', 'root_baz']);
-    });
-    it('should modify an allOf field when the defaults are set', async () => {
-      const schema: RJSFSchema = {
-        properties: {
-          all_of_field: {
-            allOf: [
-              {
-                properties: {
-                  first: {
-                    type: 'string',
-                  },
-                },
-              },
-              {
-                properties: {
-                  second: {
-                    type: 'string',
-                  },
-                },
-              },
-            ],
-            default: {
-              second: 'second!',
-            },
-          },
-        },
-        type: 'object',
-      };
-
-      const { node, onChange } = createFormComponent({
-        schema,
-      });
-
-      const secondInputID = '#root_all_of_field_second';
-      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'second!');
-
-      await user.clear(node.querySelector(secondInputID)!);
-      await user.type(node.querySelector(secondInputID)!, 'changed!');
-
-      expectToHaveBeenCalledWithFormData(
-        onChange,
-        {
-          all_of_field: {
-            second: 'changed!',
-          },
-        },
-        'root_all_of_field_second',
-      );
-
-      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'changed!');
-    });
-    it('should modify an oneOf field when the defaults are set', async () => {
-      const schema: RJSFSchema = {
-        properties: {
-          one_of_field: {
-            oneOf: [
-              {
-                properties: {
-                  first: {
-                    type: 'string',
-                  },
-                },
-              },
-              {
-                properties: {
-                  second: {
-                    type: 'string',
-                  },
-                },
-              },
-            ],
-            default: {
-              second: 'second!',
-            },
-          },
-        },
-        type: 'object',
-      };
-
-      const { node, onChange } = createFormComponent({
-        schema,
-      });
-
-      const secondInputID = '#root_one_of_field_second';
-      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'second!');
-
-      await user.clear(node.querySelector(secondInputID)!);
-      await user.type(node.querySelector(secondInputID)!, 'changed!');
-
-      expectToHaveBeenCalledWithFormData(
-        onChange,
-        {
-          one_of_field: {
-            second: 'changed!',
-          },
-        },
-        'root_one_of_field_second',
-      );
-
-      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'changed!');
-    });
-    it('should modify an anyOf field when the defaults are set', async () => {
-      const schema: RJSFSchema = {
-        properties: {
-          any_of_field: {
-            anyOf: [
-              {
-                properties: {
-                  first: {
-                    type: 'string',
-                  },
-                },
-              },
-              {
-                properties: {
-                  second: {
-                    type: 'string',
-                  },
-                },
-              },
-            ],
-            default: {
-              second: 'second!',
-            },
-          },
-        },
-        type: 'object',
-      };
-
-      const { node, onChange } = createFormComponent({
-        schema,
-      });
-
-      const secondInputID = '#root_any_of_field_second';
-      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'second!');
-
-      await user.clear(node.querySelector(secondInputID)!);
-      await user.type(node.querySelector(secondInputID)!, 'changed!');
-
-      expectToHaveBeenCalledWithFormData(
-        onChange,
-        {
-          any_of_field: {
-            second: 'changed!',
-          },
-        },
-        'root_any_of_field_second',
-      );
-
-      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'changed!');
-    });
-    it('should restore defaults when switching from null back to object option in oneOf', async () => {
-      // This test verifies that when switching from a null oneOf option back to an object option,
-      // the defaults are correctly restored. Without the fix, the form would show empty/undefined values.
-      const schema: RJSFSchema = {
-        type: 'object',
-        title: 'Configuration',
-        oneOf: [
-          {
-            title: 'Default Configuration',
-            type: 'object',
-            properties: {
-              types: { const: 'default', title: 'Types' },
-              content: { type: 'string', title: 'Content' },
-            },
-            required: ['types'],
-          },
-          {
-            title: 'Advanced Configuration',
-            type: 'object',
-            properties: {
-              types: { const: 'advanced', title: 'Types' },
-              content: { type: 'string', title: 'Content' },
-            },
-            required: ['types'],
-          },
-          { title: 'No Configuration', type: 'null' },
-        ],
-        default: { types: 'advanced', content: 'placeholder' },
-      };
-
-      const onChangeCalls: { event: IChangeEvent; id?: string }[] = [];
-
-      const { node } = createFormComponent({
-        schema,
-        onChange: (event: IChangeEvent, id?: string) => onChangeCalls.push({ event, id }),
-        experimental_defaultFormStateBehavior: { emptyObjectFields: 'populateAllDefaults' },
-      });
-
-      // Should start with "Advanced Configuration" (index 1) based on default
-      const oneOfSelect = node.querySelector<HTMLSelectElement>('#root__oneof_select');
-      expect(oneOfSelect).toBeInTheDocument();
-      expect(oneOfSelect!.value).toEqual('1');
-
-      // The content field should have the default value
-      let contentInput = node.querySelector<HTMLInputElement>('#root_content');
-      expect(contentInput).toBeInTheDocument();
-      expect(contentInput!.value).toEqual('placeholder');
-
-      // Switch to "No Configuration" (null option, index 2)
-      await user.selectOptions(oneOfSelect!, '2');
-
-      // Verify we're now on null option - content field should not exist
-      expect(node.querySelector<HTMLSelectElement>('#root__oneof_select')!.value).toEqual('2');
-      expect(node.querySelector('#root_content')).not.toBeInTheDocument();
-
-      // Switch back to "Advanced Configuration" (index 1)
-      await user.selectOptions(node.querySelector<HTMLSelectElement>('#root__oneof_select')!, '1');
-
-      // The content field should be restored with defaults
-      expect(node.querySelector<HTMLSelectElement>('#root__oneof_select')!.value).toEqual('1');
-      contentInput = node.querySelector<HTMLInputElement>('#root_content');
-      expect(contentInput).toBeInTheDocument();
-      // BUG: Without the fix, this would be empty string or undefined
-      expect(contentInput!.value).toEqual('placeholder');
-
-      // Also verify the final formData has correct values
-      const lastFormData = onChangeCalls[onChangeCalls.length - 1].event.formData;
-      expect(lastFormData.types).toEqual('advanced');
-      expect(lastFormData.content).toEqual('placeholder');
-    });
-    it('should allow switching to null option in oneOf', async () => {
-      // This test verifies that switching to a null option in oneOf works correctly.
-      // Without the fix, the form would revert back to the previous option.
-      // NOTE: This bug only manifests in controlled forms where parent updates formData prop.
-      const schema: RJSFSchema = {
-        type: 'object',
-        title: 'Configuration',
-        oneOf: [
-          {
-            title: 'Default Configuration',
-            type: 'object',
-            properties: {
-              types: { const: 'default', title: 'Types' },
-              content: { type: 'string', title: 'Content' },
-            },
-            required: ['types'],
-          },
-          {
-            title: 'Advanced Configuration',
-            type: 'object',
-            properties: {
-              types: { const: 'advanced', title: 'Types' },
-              content: { type: 'string', title: 'Content' },
-            },
-            required: ['types'],
-          },
-          { title: 'No Configuration', type: 'null' },
-        ],
-        default: { types: 'advanced', content: 'placeholder' },
-      };
-
-      const onChangeCalls: { event: IChangeEvent; id?: string }[] = [];
-      let currentFormData: unknown = undefined;
-
-      const { node, rerender } = createFormComponent({
-        ref: createRef(),
-        schema,
-        onChange: (event: IChangeEvent, id?: string) => {
-          onChangeCalls.push({ event, id });
-          currentFormData = event.formData;
-        },
-        experimental_defaultFormStateBehavior: { emptyObjectFields: 'populateAllDefaults' },
-      });
-
-      // Should start with "Advanced Configuration" (index 1)
-      expect(node.querySelector<HTMLSelectElement>('#root__oneof_select')!.value).toEqual('1');
-      expect(node.querySelector('#root_content')).toBeInTheDocument();
-
-      // Switch to "No Configuration" (null option, index 2)
-      await user.selectOptions(node.querySelector<HTMLSelectElement>('#root__oneof_select')!, '2');
-
-      // Simulate controlled form behavior by re-rendering with new formData
-      rerender({
-        ref: createRef(),
-        schema,
-        formData: currentFormData,
-        experimental_defaultFormStateBehavior: { emptyObjectFields: 'populateAllDefaults' },
-      });
-
-      // BUG: Without the fix, the form would revert back to index 1
-      expect(node.querySelector<HTMLSelectElement>('#root__oneof_select')!.value).toEqual('2');
-      // Content field should not exist for null option
-      expect(node.querySelector('#root_content')).not.toBeInTheDocument();
-
-      // Verify formData is null or undefined (both valid for null option)
-      const lastFormData = onChangeCalls[onChangeCalls.length - 1].event.formData;
-      expect(lastFormData == null).toBe(true);
-    });
-    it('Should modify anyOf definition references when the defaults are set.', async () => {
-      const schema: RJSFSchema = {
-        definitions: {
-          option1: {
-            properties: {
-              first: {
-                type: 'string',
-              },
-            },
-          },
-          option2: {
-            properties: {
-              second: {
-                type: 'string',
-              },
-            },
-          },
-        },
-        properties: {
-          any_of_field: {
-            anyOf: [
-              {
-                $ref: '#/definitions/option1',
-              },
-              {
-                $ref: '#/definitions/option2',
-              },
-            ],
-            default: {
-              second: 'second!',
-            },
-          },
-        },
-        type: 'object',
-      };
-
-      const { node, onChange } = createFormComponent({
-        schema,
-      });
-
-      const secondInputID = '#root_any_of_field_second';
-      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'second!');
-
-      await user.clear(node.querySelector(secondInputID)!);
-      await user.type(node.querySelector(secondInputID)!, 'changed!');
-
-      expectToHaveBeenCalledWithFormData(
-        onChange,
-        {
-          any_of_field: {
-            second: 'changed!',
-          },
-        },
-        'root_any_of_field_second',
-      );
-
-      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'changed!');
-    });
-    it('Should modify oneOf object with references when the defaults are set.', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        $defs: {
-          protocol: {
-            type: 'string',
-            enum: ['fast', 'balanced', 'stringent'],
-            default: 'fast',
-          },
-        },
-        oneOf: [
-          {
-            properties: {
-              protocol: {
-                $ref: '#/$defs/protocol',
-              },
-            },
-          },
-          {
-            properties: {
-              something: {
-                type: 'number',
-              },
-            },
-          },
-        ],
-      };
-
-      const { node, onChange } = createFormComponent({
-        schema,
-      });
-
-      const protocolInputID = '#root_protocol';
-      expect(node.querySelector(protocolInputID)).toHaveValue('0');
-
-      await user.selectOptions(node.querySelector(protocolInputID)!, '1');
-
-      expectToHaveBeenCalledWithFormData(
-        onChange,
-        {
-          protocol: 'balanced',
-        },
-        'root_protocol',
-      );
-
-      expect(node.querySelector(protocolInputID)).toHaveValue('1');
-    });
-    describe('Should modify oneOf radio button when the defaults are set.', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          a: {
-            type: ['boolean', 'null'],
-            default: null,
-            oneOf: [
-              {
-                const: false,
-                title: 'No',
-              },
-              {
-                const: null,
-                title: 'N/A',
-              },
-            ],
-          },
-        },
-        allOf: [
-          {
-            if: {
-              required: ['a'],
-              properties: {
-                a: {
-                  const: false,
-                },
-              },
-            },
-            then: {
-              required: ['b'],
-              properties: {
-                b: {
-                  type: 'string',
-                },
-              },
-            },
-          },
-        ],
-      };
-
-      const uiSchema: UiSchema = {
-        a: {
-          'ui:widget': 'radio',
-          'ui:label': false,
-        },
-      };
-      const notApplicableInputID = '#root_a-1';
-      const NoInputID = '#root_a-0';
-
-      it('Test with default constAsDefaults', async () => {
-        const { node, onChange } = createFormComponent({
-          schema,
-          uiSchema,
-        });
-
-        expect(node.querySelector(notApplicableInputID)).toBeChecked();
-
-        await user.click(node.querySelector(NoInputID)!);
-
-        expectToHaveBeenCalledWithFormData(onChange, { a: false }, 'root_a');
-
-        expect(node.querySelector(NoInputID)).toBeChecked();
-        expect(node.querySelector(notApplicableInputID)).not.toBeChecked();
-        expect(node.querySelector('#root_b')).toBeInTheDocument();
-      });
-      it('Test with constAsDefaults set to "never"', async () => {
-        const { node, onChange } = createFormComponent({
-          schema,
-          uiSchema,
-          experimental_defaultFormStateBehavior: {
-            constAsDefaults: 'never',
-          },
-        });
-
-        expect(node.querySelector(notApplicableInputID)).toBeChecked();
-
-        await user.click(node.querySelector(NoInputID)!);
-
-        expectToHaveBeenCalledWithFormData(onChange, { a: false }, 'root_a');
-
-        expect(node.querySelector(NoInputID)).toBeChecked();
-        expect(node.querySelector(notApplicableInputID)).not.toBeChecked();
-        expect(node.querySelector('#root_b')).toBeInTheDocument();
-      });
-    });
-  });
-
-  describe('Blur handler', () => {
-    it('should call provided blur handler on form input blur event', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          foo: {
-            type: 'string',
-          },
-        },
-      };
-      const formData = {
-        foo: '',
-      };
-      const onBlur = vi.fn();
-      const { node } = createFormComponent({ schema, formData, onBlur });
-
-      const input = node.querySelector('[type=text]')!;
-      await user.type(input, 'new');
-      await user.tab();
-
-      expect(onBlur).toHaveBeenLastCalledWith(input.id, 'new');
-    });
-  });
-
-  describe('Focus handler', () => {
-    it('should call provided focus handler on form input focus event', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          foo: {
-            type: 'string',
-          },
-        },
-      };
-      const formData = {
-        foo: 'new',
-      };
-      const onFocus = vi.fn();
-      const { node } = createFormComponent({ schema, formData, onFocus });
-
-      const input = node.querySelector('[type=text]')!;
-      await user.click(input);
-
-      expect(onFocus).toHaveBeenLastCalledWith(input.id, 'new');
-    });
-  });
-
-  describe('Error handler', () => {
-    it('should call provided error handler on validation errors', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          foo: {
-            type: 'string',
-            minLength: 1,
-          },
-        },
-      };
-      const formData = {
-        foo: '',
-      };
-      const { node, onError } = createFormComponent({ schema, formData });
-
-      await submitForm(node, user);
-
-      expect(onError).toHaveBeenCalledTimes(1);
-    });
-  });
-
-  describe('Required and optional fields', () => {
-    const schema: RJSFSchema = {
-      definitions: {
-        address: {
-          type: 'object',
-          properties: {
-            street_address: {
-              type: 'string',
-            },
-            city: {
-              type: 'string',
-            },
-            state: {
-              type: 'string',
-            },
-          },
-          required: ['street_address', 'city', 'state'],
-        },
-      },
-      type: 'object',
-      properties: {
-        billing_address: {
-          title: 'Billing address',
-          $ref: '#/definitions/address',
-        },
-        shipping_address: {
-          title: 'Shipping address',
-          $ref: '#/definitions/address',
-        },
-      },
-      required: ['shipping_address'],
-    };
-    it('Errors when shipping address is not filled out, billing address is not needed', async () => {
-      const { node, onChange, onError } = createFormComponent({ schema });
-      expectToHaveBeenCalledWithFormData(onChange, { shipping_address: {} });
-      // forceFireEvent=true: clicking the submit button focuses it, blurring the
-      // currently focused field and firing onChange which may mutate formData before
-      // the submit handler runs. fireEvent.submit bypasses that side-effect chain.
-      await submitForm(node, user, true);
-      expect(onError).toHaveBeenLastCalledWith([
-        {
-          message: "must have required property 'street_address'",
-          name: 'required',
-          params: { missingProperty: 'street_address' },
-          property: '.shipping_address.street_address',
-          schemaPath: '#/definitions/address/required',
-          stack: "must have required property 'street_address'",
-          title: '',
-        },
-        {
-          message: "must have required property 'city'",
-          name: 'required',
-          params: { missingProperty: 'city' },
-          property: '.shipping_address.city',
-          schemaPath: '#/definitions/address/required',
-          stack: "must have required property 'city'",
-          title: '',
-        },
-        {
-          message: "must have required property 'state'",
-          name: 'required',
-          params: { missingProperty: 'state' },
-          property: '.shipping_address.state',
-          schemaPath: '#/definitions/address/required',
-          stack: "must have required property 'state'",
-          title: '',
-        },
-      ]);
-    });
-    it('Submits when shipping address is filled out, billing address is not needed', async () => {
-      const { node, onSubmit } = createFormComponent({
-        schema,
-        formData: {
-          shipping_address: {
-            street_address: '21, Jump Street',
-            city: 'Babel',
-            state: 'Neverland',
-          },
-        },
-      });
-      // forceFireEvent=true: clicking the submit button focuses it, blurring the
-      // currently focused field and firing onChange which may mutate formData before
-      // the submit handler runs. fireEvent.submit bypasses that side-effect chain.
-      await submitForm(node, user, true);
-      expectToHaveBeenCalledWithFormData(
-        onSubmit,
-        {
-          shipping_address: {
-            street_address: '21, Jump Street',
-            city: 'Babel',
-            state: 'Neverland',
-          },
-        },
-        true,
-      );
-    });
-  });
-
-  describe('Default form state behavior flag', () => {
-    const schema: RJSFSchema = {
-      type: 'object',
-      properties: {
-        albums: {
-          type: 'array',
-          items: { type: 'string' },
-          title: 'Album Titles',
-          minItems: 3,
-        },
-      },
-    };
-    it('Errors when minItems is set, field is required, and minimum number of items are not present with IgnoreMinItemsUnlessRequired flag set', async () => {
-      const { node, onError } = createFormComponent({
-        schema: { ...schema, required: ['albums'] },
-        formData: {
-          albums: ['Until We Have Faces'],
-        },
-        experimental_defaultFormStateBehavior: { arrayMinItems: { populate: 'requiredOnly' } },
-      });
-      await submitForm(node, user);
-      expect(onError).toHaveBeenLastCalledWith([
-        {
-          message: 'must NOT have fewer than 3 items',
-          name: 'minItems',
-          params: { limit: 3 },
-          property: '.albums',
-          schemaPath: '#/properties/albums/minItems',
-          stack: "'Album Titles' must NOT have fewer than 3 items",
-          title: 'Album Titles',
-        },
-      ]);
-    });
-    it('Submits when minItems is set, field is not required, and no items are present with IgnoreMinItemsUnlessRequired flag set', async () => {
-      const { node, onSubmit } = createFormComponent({
-        schema,
-        formData: {},
-        experimental_defaultFormStateBehavior: { arrayMinItems: { populate: 'requiredOnly' } },
-      });
-      await submitForm(node, user);
-      expectToHaveBeenCalledWithFormData(onSubmit, {}, true);
-    });
-  });
-
-  describe('Schema and external formData updates', () => {
-    let rerender: RerenderType;
-    let onChangeProp: Mock;
-    let formProps: NoValFormProps;
-
-    beforeEach(() => {
-      formProps = {
-        ref: createRef(),
-        schema: {
-          type: 'string',
-          default: 'foobar',
-        },
-        formData: 'some value',
-      };
-      const { rerender: rerenderFn, onChange } = createFormComponent(formProps);
-      onChangeProp = onChange;
-      rerender = rerenderFn;
-    });
-
-    describe('when the form data is set to null', () => {
-      beforeEach(() =>
-        rerender({
-          ...formProps,
-          formData: null,
-        }),
-      );
-
-      it('should call onChange', () => {
-        expect(onChangeProp).toHaveBeenCalledTimes(1);
-        expect(onChangeProp).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            edit: true,
-            errorSchema: {},
-            errors: [],
-            formData: 'foobar',
-            fieldPathId: { $id: 'root', path: [] },
-            schema: formProps.schema,
-            uiSchema: {},
-            schemaUtils: expect.any(Object),
-          }),
-        );
-      });
-    });
-
-    describe('when the schema default is changed but formData is not changed', () => {
-      const newSchema: RJSFSchema = {
-        type: 'string',
-        default: 'the new default',
-      };
-
-      beforeEach(() =>
-        rerender({
-          ...formProps,
-          schema: newSchema,
-          formData: 'some value',
-        }),
-      );
-
-      it('should not call onChange', () => {
-        expect(onChangeProp).not.toHaveBeenCalled();
-      });
-    });
-
-    describe('when the schema default is changed and formData is changed', () => {
-      const newSchema: RJSFSchema = {
-        type: 'string',
-        default: 'the new default',
-      };
-
-      beforeEach(() =>
-        rerender({
-          ...formProps,
-          schema: newSchema,
-          formData: 'something else',
-        }),
-      );
-
-      it('should not call onChange', () => {
-        expect(onChangeProp).not.toHaveBeenCalled();
-      });
-    });
-
-    describe('when the schema default is changed and formData is nulled', () => {
-      const newSchema: RJSFSchema = {
-        type: 'string',
-        default: 'the new default',
-      };
-
-      beforeEach(() =>
-        rerender({
-          ...formProps,
-          schema: newSchema,
-          formData: null,
-        }),
-      );
-
-      it('should call onChange', () => {
-        expect(onChangeProp).toHaveBeenCalledTimes(1);
-        expect(onChangeProp).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            schema: newSchema,
-            formData: 'the new default',
-          }),
-        );
-      });
-    });
-
-    describe('when the onChange prop sets formData to a falsey value', () => {
-      function TestForm(props: { falseyValue: any }) {
-        const [formData, setFormData] = useState<any>({});
-
-        const onChange = useCallback(() => {
-          setFormData(props.falseyValue);
-        }, [props]);
-        const schema: RJSFSchema = {
-          type: 'object',
-          properties: {
-            value: {
-              type: 'string',
-            },
-          },
-        };
-        return <Form onChange={onChange} schema={schema} formData={formData} validator={validator} />;
-      }
-
-      const falseyValues = [0, false, null, undefined, NaN];
-
-      falseyValues.forEach((falseyValue) => {
-        it("Should not crash due to 'Maximum call stack size exceeded...'", () => {
-          // It is expected that this will throw an error due to non-matching propTypes,
-          // so the error message needs to be inspected
-          try {
-            renderNode(TestForm, { falseyValue });
-          } catch (e) {
-            const { message } = e as Error;
-            expect(message).not.toEqual('Maximum call stack size exceeded');
-          }
-        });
-      });
-    });
-  });
-
-  describe('External formData updates', () => {
-    describe('root level', () => {
-      const formProps: Omit<FormProps, 'validator'> = {
-        ref: createRef(),
-        schema: { type: 'string' },
-      };
-
-      it('should call submit handler with new formData prop value', async () => {
-        const { node, onSubmit, rerender } = createFormComponent(formProps);
-
-        rerender({
-          ...formProps,
-          onSubmit,
-          formData: 'yo',
-        });
-        await submitForm(node, user);
-        expectToHaveBeenCalledWithFormData(onSubmit, 'yo', true);
-      });
-
-      it('should validate formData when the schema is updated', async () => {
-        const { node, onError, rerender } = createFormComponent(formProps);
-
-        rerender({
-          ...formProps,
-          onError,
-          formData: 'yo',
-          schema: { type: 'number' },
-        });
-        await submitForm(node, user);
-        expect(onError).toHaveBeenLastCalledWith([
-          {
-            message: 'must be number',
-            name: 'type',
-            params: { type: 'number' },
-            property: '',
-            schemaPath: '#/type',
-            stack: 'must be number',
-            title: '',
-          },
-        ]);
-      });
-    });
-
-    describe('object level', () => {
-      it('should call submit handler with new formData prop value', async () => {
-        const formProps: Omit<FormProps, 'validator'> = {
-          ref: createRef(),
-          schema: { type: 'object', properties: { foo: { type: 'string' } } },
-        };
-        const { onSubmit, node, rerender } = createFormComponent(formProps);
-
-        rerender({
-          ...formProps,
-          onSubmit,
-          formData: { foo: 'yo' },
-        });
-
-        await submitForm(node, user);
-        expectToHaveBeenCalledWithFormData(onSubmit, { foo: 'yo' }, true);
-      });
-    });
-
-    describe('array level', () => {
-      it('should call submit handler with new formData prop value', async () => {
-        const schema: RJSFSchema = {
-          type: 'array',
-          items: {
-            type: 'string',
-          },
-        };
-        const { node, onSubmit, rerender } = createFormComponent({ ref: createRef(), schema });
-
-        rerender({
-          schema,
-          onSubmit,
-          formData: ['yo'],
-        });
-
-        await submitForm(node, user);
-        expectToHaveBeenCalledWithFormData(onSubmit, ['yo'], true);
-      });
-    });
-  });
-
-  describe('Internal formData updates', () => {
-    it('root', async () => {
-      const { node, onChange } = createFormComponent({
-        ref: createRef(),
-        schema: { type: 'string' },
-      });
-
-      await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'yo');
-
-      expectToHaveBeenCalledWithFormData(onChange, 'yo', 'root');
-    });
-    it('object', async () => {
-      const { node, onChange } = createFormComponent({
-        ref: createRef(),
-        schema: {
-          type: 'object',
-          properties: {
-            foo: {
-              type: 'string',
-            },
-          },
-        },
-      });
-
-      await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'yo');
-
-      expectToHaveBeenCalledWithFormData(onChange, { foo: 'yo' }, 'root_foo');
-    });
-    it('array of strings', async () => {
-      const schema: RJSFSchema = {
-        type: 'array',
-        items: {
-          type: 'string',
-        },
-      };
-      const { node, onChange } = createFormComponent({ ref: createRef(), schema });
-
-      await user.click(node.querySelector('.rjsf-array-item-add button')!);
-
-      await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'yo');
-      expectToHaveBeenCalledWithFormData(onChange, ['yo'], 'root_0');
-    });
-    it('array of objects', async () => {
-      const schema: RJSFSchema = {
-        type: 'array',
-        items: {
-          type: 'object',
-          properties: {
-            name: { type: 'string' },
-          },
-        },
-      };
-      const { node, onChange } = createFormComponent({ ref: createRef(), schema });
-
-      await user.click(node.querySelector('.rjsf-array-item-add button')!);
-
-      await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'yo');
-
-      expectToHaveBeenCalledWithFormData(onChange, [{ name: 'yo' }], 'root_0_name');
-    });
-    it('dependency with array of objects', async () => {
-      const schema: RJSFSchema = {
-        definitions: {},
-        type: 'object',
-        properties: {
-          show: {
-            type: 'boolean',
-          },
-        },
-        dependencies: {
-          show: {
-            oneOf: [
-              {
-                properties: {
-                  show: {
-                    const: true,
-                  },
-                  participants: {
-                    type: 'array',
-                    items: {
-                      type: 'object',
-                      properties: {
-                        name: {
-                          type: 'string',
-                        },
-                      },
-                    },
-                  },
-                },
-              },
-            ],
-          },
-        },
-      };
-      const { node, onChange } = createFormComponent({ ref: createRef(), schema });
-
-      const checkbox = node.querySelector('[type=checkbox]');
-      await user.click(checkbox!);
-
-      await user.click(node.querySelector('.rjsf-array-item-add button')!);
-
-      await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'yo');
-
-      expectToHaveBeenCalledWithFormData(
-        onChange,
-        {
-          show: true,
-          participants: [{ name: 'yo' }],
-        },
-        'root_participants_0_name',
-      );
-    });
-  });
-
-  describe('Error contextualization', () => {
-    describe('on form state updated', () => {
-      const schema: RJSFSchema = {
-        type: 'string',
-        minLength: 8,
-      };
-
-      describe('Lazy validation', () => {
-        it('should not update the errorSchema when the formData changes', async () => {
-          const { node, onChange } = createFormComponent({ schema });
-
-          await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'short');
-          expect(onChange).toHaveBeenLastCalledWith(expect.objectContaining({ errorSchema: {} }), 'root');
-        });
-
-        it('should not denote an error in the field', async () => {
-          const { node } = createFormComponent({ schema });
-
-          await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'short');
-
-          expect(node.querySelectorAll('.rjsf-field-error')).toHaveLength(0);
-        });
-
-        it("should clean contextualized errors up when they're fixed", async () => {
-          const altSchema: RJSFSchema = {
-            type: 'object',
-            properties: {
-              field1: { type: 'string', minLength: 8 },
-              field2: { type: 'string', minLength: 8 },
-            },
-          };
-          const { node } = createFormComponent({
-            schema: altSchema,
-            formData: {
-              field1: 'short',
-              field2: 'short',
-            },
-          });
-
-          await submitForm(node, user);
-
-          // Fix the first field
-          await user.clear(node.querySelectorAll('input[type=text]')[0]);
-          await user.type(node.querySelectorAll('input[type=text]')[0], 'fixed error');
-          await submitForm(node, user);
-
-          expect(node.querySelectorAll('.rjsf-field-error')).toHaveLength(1);
-
-          // Fix the second field
-          await user.clear(node.querySelectorAll('input[type=text]')[1]);
-          await user.type(node.querySelectorAll('input[type=text]')[1], 'fixed error too');
-          await submitForm(node, user);
-
-          // No error remaining, shouldn't throw.
-          await submitForm(node, user);
-
-          expect(node.querySelectorAll('.rjsf-field-error')).toHaveLength(0);
-        });
-
-        it('should only clear the error of the field that changed inside an array item (#5197)', async () => {
-          const altSchema: RJSFSchema = {
-            type: 'object',
-            properties: {
-              foo: { type: 'string' },
-              bar: { type: 'string' },
-              baz: {
-                type: 'array',
-                items: {
-                  type: 'object',
-                  properties: {
-                    qux: { type: 'string' },
-                    corge: { type: 'string' },
-                  },
-                  required: ['qux', 'corge'],
-                },
-              },
-            },
-            required: ['foo', 'bar'],
-          };
-
-          // The form has to be controlled, since the errors are cleared by comparing the incoming
-          // `formData` prop against the previous one.
-          function Controlled() {
-            const [formData, setFormData] = useState<any>({ baz: [{}] });
-            return (
-              <Form
-                schema={altSchema}
-                validator={validator}
-                formData={formData}
-                noHtml5Validate
-                onChange={(e) => setFormData(e.formData)}
-              />
-            );
-          }
-
-          const { container } = render(<Controlled />);
-          const node = container.firstElementChild!;
-          const shownErrors = () => Array.from(node.querySelectorAll('.error-detail')).map((e) => e.textContent);
-
-          await submitForm(node, user);
-          expect(shownErrors()).toEqual([
-            "must have required property 'foo'",
-            "must have required property 'bar'",
-            "must have required property 'qux'",
-            "must have required property 'corge'",
-          ]);
-
-          // A top-level field clears only itself, which already worked
-          await user.type(node.querySelector('#root_foo')!, 'a');
-          expect(shownErrors()).toEqual([
-            "must have required property 'bar'",
-            "must have required property 'qux'",
-            "must have required property 'corge'",
-          ]);
-
-          // and so should a field inside an array item
-          await user.type(node.querySelector('#root_baz_0_qux')!, 'a');
-          expect(shownErrors()).toEqual(["must have required property 'bar'", "must have required property 'corge'"]);
-        });
-
-        it('should clear the error of a field whose name contains a dot', async () => {
-          const altSchema: RJSFSchema = {
-            type: 'object',
-            properties: {
-              'foo.bar': { type: 'string' },
-              baz: { type: 'string' },
-            },
-            required: ['foo.bar', 'baz'],
-          };
-          const formRef = createRef<Form>();
-
-          function Controlled() {
-            const [formData, setFormData] = useState<any>({});
-            return (
-              <Form
-                ref={formRef}
-                schema={altSchema}
-                validator={validator}
-                formData={formData}
-                noHtml5Validate
-                onChange={(e) => setFormData(e.formData)}
-              />
-            );
-          }
-
-          const { container } = render(<Controlled />);
-          const node = container.firstElementChild!;
-
-          await submitForm(node, user);
-          // `toErrorSchema` runs the property name through `toPath`, so the error of a name holding a dot lands at
-          // the path that name spells out, not under the name itself
-          expect(formRef.current!.state.errorSchema).toEqual({
-            foo: { bar: { __errors: ["must have required property 'foo.bar'"] } },
-            baz: { __errors: ["must have required property 'baz'"] },
-          });
-
-          // Clearing has to reach the same place, and leave the field that was not touched alone
-          await user.type(node.querySelector('[id="root_foo.bar"]')!, 'a');
-          expect(formRef.current!.state.errorSchema).toEqual({
-            foo: { bar: undefined },
-            baz: { __errors: ["must have required property 'baz'"] },
-          });
-        });
-
-        it('should only clear the error of the field that changed under a name containing a dot', async () => {
-          const altSchema: RJSFSchema = {
-            type: 'object',
-            properties: {
-              'has.dot': {
-                type: 'object',
-                properties: {
-                  inner: { type: 'string' },
-                  other: { type: 'string' },
-                },
-                required: ['inner', 'other'],
-              },
-            },
-          };
-          const formRef = createRef<Form>();
-
-          function Controlled() {
-            const [formData, setFormData] = useState<any>({ 'has.dot': {} });
-            return (
-              <Form
-                ref={formRef}
-                schema={altSchema}
-                validator={validator}
-                formData={formData}
-                noHtml5Validate
-                onChange={(e) => setFormData(e.formData)}
-              />
-            );
-          }
-
-          const { container } = render(<Controlled />);
-          const node = container.firstElementChild!;
-
-          await submitForm(node, user);
-          // The name is spelled out as a path here too, so the errors of the two fields it holds sit side by side
-          // under it and clearing one has to leave the other where it is
-          expect(formRef.current!.state.errorSchema).toEqual({
-            has: {
-              dot: {
-                inner: { __errors: ["must have required property 'inner'"] },
-                other: { __errors: ["must have required property 'other'"] },
-              },
-            },
-          });
-
-          await user.type(node.querySelector('[id="root_has.dot_inner"]')!, 'a');
-          expect(formRef.current!.state.errorSchema).toEqual({
-            has: {
-              dot: {
-                inner: undefined,
-                other: { __errors: ["must have required property 'other'"] },
-              },
-            },
-          });
-        });
-
-        it('should clear the error of the container holding the field that changed', async () => {
-          const altSchema: RJSFSchema = {
-            type: 'object',
-            properties: {
-              tags: { type: 'array', uniqueItems: true, items: { type: 'string' } },
-            },
-          };
-
-          function Controlled() {
-            const [formData, setFormData] = useState<any>({ tags: ['a', 'a'] });
-            return (
-              <Form
-                schema={altSchema}
-                validator={validator}
-                formData={formData}
-                noHtml5Validate
-                onChange={(e) => setFormData(e.formData)}
-              />
-            );
-          }
-
-          const { container } = render(<Controlled />);
-          const node = container.firstElementChild!;
-          const shownErrors = () => Array.from(node.querySelectorAll('.error-detail')).map((e) => e.textContent);
-
-          await submitForm(node, user);
-          expect(shownErrors()).toEqual(['must NOT have duplicate items (items ## 1 and 0 are identical)']);
-
-          // The error belongs to the array rather than to the item that was edited, and the array changed too
-          await user.type(node.querySelectorAll('input[type=text]')[1], 'b');
-          expect(shownErrors()).toEqual([]);
-        });
-
-        it('should keep the errors that a reorder of the array items left standing', async () => {
-          const altSchema: RJSFSchema = {
-            type: 'object',
-            properties: {
-              baz: {
-                type: 'array',
-                items: {
-                  type: 'object',
-                  properties: {
-                    qux: { type: 'string' },
-                    corge: { type: 'string' },
-                  },
-                  required: ['qux', 'corge'],
-                },
-              },
-            },
-          };
-
-          function Controlled() {
-            const [formData, setFormData] = useState<any>({ baz: [{}, { corge: 'z' }] });
-            return (
-              <Form
-                schema={altSchema}
-                validator={validator}
-                formData={formData}
-                noHtml5Validate
-                onChange={(e) => setFormData(e.formData)}
-              />
-            );
-          }
-
-          const { container } = render(<Controlled />);
-          const node = container.firstElementChild!;
-          const shownErrors = () => Array.from(node.querySelectorAll('.error-detail')).map((e) => e.textContent);
-
-          await submitForm(node, user);
-          expect(shownErrors()).toEqual([
-            "must have required property 'qux'",
-            "must have required property 'corge'",
-            "must have required property 'qux'",
-          ]);
-
-          // Moving the first item down leaves neither item holding a `qux`, so both of those errors are still the
-          // errors of the values sitting at those paths and only the `corge` that moved loses its own
-          await user.click(node.querySelector<HTMLButtonElement>('.rjsf-array-item-move-down')!);
-          expect(shownErrors()).toEqual(["must have required property 'qux'", "must have required property 'qux'"]);
-        });
-      });
-
-      describe('Live validation', () => {
-        it('should update the errorSchema when the formData changes', async () => {
-          const { node, onChange } = createFormComponent({
-            schema,
-            liveValidate: true,
-          });
-
-          await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'short');
-
-          expect(onChange).toHaveBeenLastCalledWith(
-            expect.objectContaining({
-              errorSchema: {
-                __errors: ['must NOT have fewer than 8 characters'],
-              },
-            }),
-            'root',
-          );
-        });
-
-        it('should denote the new error in the field', async () => {
-          const { node } = createFormComponent({
-            schema,
-            liveValidate: true,
-          });
-
-          await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'short');
-
-          expect(node.querySelectorAll('.rjsf-field-error')).toHaveLength(1);
-          expect(node.querySelector('.rjsf-field-string .error-detail')).toHaveTextContent(
-            'must NOT have fewer than 8 characters',
-          );
-        });
-      });
-
-      describe('Disable validation onChange event', () => {
-        it('should not update errorSchema when the formData changes', async () => {
-          const { node, onChange } = createFormComponent({
-            schema,
-            noValidate: true,
-            liveValidate: true,
-          });
-
-          await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'short');
-
-          expect(onChange).toHaveBeenLastCalledWith(expect.objectContaining({ errorSchema: {} }), 'root');
-        });
-      });
-
-      describe('Disable validation onSubmit event', () => {
-        it('should not update errorSchema when the formData changes', async () => {
-          const { node, onSubmit } = createFormComponent({
-            schema,
-            noValidate: true,
-          });
-
-          await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'short');
-          await submitForm(node, user);
-
-          expect(onSubmit).toHaveBeenLastCalledWith(
-            expect.objectContaining({ errorSchema: {} }),
-            expect.objectContaining({ type: 'submit' }),
-          );
-        });
-      });
-    });
-
-    describe('on form submitted', () => {
-      const schema: RJSFSchema = {
-        type: 'string',
-        minLength: 8,
-      };
-
-      it('should call the onError handler and focus on the error', async () => {
-        const onError = vi.fn();
-        const { node } = createFormComponent({
-          schema,
-          onError,
-          focusOnFirstError: true,
-        });
-
-        const input = node.querySelector<HTMLInputElement>('input[type=text]')!;
-        await user.type(input, 'short');
-        const focusSpy = vi.fn();
-        // Since programmatically triggering focus does not call onFocus, change the focus method to a spy
-        // Set up AFTER user.type so typing works (focus spy would prevent element from receiving focus)
-        Object.defineProperty(input, 'focus', {
-          configurable: true,
-          value: focusSpy,
-        });
-        await submitForm(node, user);
-
-        expect(onError).toHaveBeenLastCalledWith(expect.any(Array));
-        const callArg = vi.mocked(onError).mock.calls[0][0];
-        expect(callArg.length).toBe(1);
-        expect(callArg[0].message).toBe('must NOT have fewer than 8 characters');
-        expect(focusSpy).toHaveBeenCalledTimes(1);
-      });
-
-      it('should call the onError handler and call the provided focusOnFirstError callback function', async () => {
-        const consoleLogSpy = vi.spyOn(console, 'log').mockImplementation(noop);
-        const onError = vi.fn();
-
-        const focusCallback = () => {
-          // oxlint-disable-next-line no-console
-          console.log('focusCallback called');
-        };
-
-        const focusOnFirstError = vi.fn(focusCallback);
-        const { node } = createFormComponent({
-          schema,
-          onError,
-          focusOnFirstError,
-        });
-
-        const input = node.querySelector<HTMLInputElement>('input[type=text]')!;
-        await user.type(input, 'short');
-        const focusSpy = vi.fn();
-        // Since programmatically triggering focus does not call onFocus, change the focus method to a spy
-        // Set up AFTER user.type so typing works (focus spy would prevent element from receiving focus)
-        Object.defineProperty(input, 'focus', {
-          configurable: true,
-          value: focusSpy,
-        });
-        await submitForm(node, user);
-
-        expect(onError).toHaveBeenLastCalledWith(expect.any(Array));
-        const callArg = vi.mocked(onError).mock.calls[0][0];
-        expect(callArg.length).toBe(1);
-        expect(callArg[0].message).toBe('must NOT have fewer than 8 characters');
-
-        expect(focusSpy).not.toHaveBeenCalled();
-        expect(focusOnFirstError).toHaveBeenCalledTimes(1);
-        expect(consoleLogSpy).toHaveBeenCalledWith('focusCallback called');
-        consoleLogSpy.mockRestore();
-      });
-
-      it('should call the onError handler and call the provided focusOnFirstError callback function', async () => {
-        const consoleLogSpy = vi.spyOn(console, 'log').mockImplementation(noop);
-        const onError = vi.fn();
-
-        const focusCallback = () => {
-          // oxlint-disable-next-line no-console
-          console.log('focusCallback called');
-        };
-        const extraErrors = {
-          __errors: ['foo'],
-        } as ErrorSchema;
-
-        const focusOnFirstError = vi.fn(focusCallback);
-        const { node } = createFormComponent({
-          schema,
-          onError,
-          focusOnFirstError,
-          extraErrors,
-          extraErrorsBlockSubmit: true,
-        });
-
-        const input = node.querySelector<HTMLInputElement>('input[type=text]')!;
-        await user.type(input, 'valid string');
-        const focusSpy = vi.fn();
-        // Since programmatically triggering focus does not call onFocus, change the focus method to a spy
-        // Set up AFTER user.type so typing works (focus spy would prevent element from receiving focus)
-        Object.defineProperty(input, 'focus', {
-          configurable: true,
-          value: focusSpy,
-        });
-        await submitForm(node, user);
-
-        expect(onError).toHaveBeenLastCalledWith(expect.any(Array));
-        const callArg = vi.mocked(onError).mock.calls[0][0];
-        expect(callArg.length).toBe(1);
-        expect(callArg[0].message).toBe('foo');
-
-        expect(focusSpy).not.toHaveBeenCalled();
-        expect(focusOnFirstError).toHaveBeenCalledTimes(1);
-        expect(consoleLogSpy).toHaveBeenCalledWith('focusCallback called');
-        consoleLogSpy.mockRestore();
-      });
-
-      it('should reset errors and errorSchema state to initial state after correction and resubmission', async () => {
-        const { node, onError, onSubmit } = createFormComponent({
-          schema,
-        });
-
-        await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'short');
-        await submitForm(node, user);
-
-        expect(onError).toHaveBeenLastCalledWith([
-          {
-            message: 'must NOT have fewer than 8 characters',
-            name: 'minLength',
-            params: { limit: 8 },
-            property: '',
-            schemaPath: '#/minLength',
-            stack: 'must NOT have fewer than 8 characters',
-            title: '',
-          },
-        ]);
-        expect(onError).toHaveBeenCalledTimes(1);
-        onError.mockClear();
-
-        await user.clear(node.querySelector<HTMLInputElement>('input[type=text]')!);
-        await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'long enough');
-        await submitForm(node, user);
-        expect(onError).not.toHaveBeenCalled();
-        expect(onSubmit).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            errors: [],
-            errorSchema: {},
-          }),
-          expect.objectContaining({ type: 'submit' }),
-        );
-      });
-
-      it('should reset errors from UI after correction and resubmission', async () => {
-        const { node } = createFormComponent({
-          schema,
-        });
-
-        await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'short');
-        await submitForm(node, user);
-
-        const errorListHTML = '<li class="text-danger">must NOT have fewer than 8 characters</li>';
-        const errors = node.querySelectorAll('.error-detail');
-        // Check for errors attached to the field
-        expect(errors).toHaveLength(1);
-        expect(errors[0].innerHTML).toEqual(errorListHTML);
-
-        await user.clear(node.querySelector<HTMLInputElement>('input[type=text]')!);
-        await user.type(node.querySelector<HTMLInputElement>('input[type=text]')!, 'long enough');
-        await submitForm(node, user);
-        expect(node.querySelectorAll('.error-detail')).toHaveLength(0);
-      });
-    });
-
-    describe('root level, live validation', () => {
-      const formProps: Omit<FormProps, 'validator'> = {
-        liveValidate: true,
-        schema: {
-          type: 'string',
-          minLength: 8,
-        },
-        formData: 'short',
-      };
-
-      it('should reflect the contextualized error in state', async () => {
-        const { node, onError } = createFormComponent(formProps);
-        await submitForm(node, user);
-        expect(onError).toHaveBeenLastCalledWith([
-          {
-            message: 'must NOT have fewer than 8 characters',
-            name: 'minLength',
-            params: { limit: 8 },
-            property: '',
-            schemaPath: '#/minLength',
-            stack: 'must NOT have fewer than 8 characters',
-            title: '',
-          },
-        ]);
-      });
-
-      it('should denote the error in the field', async () => {
-        const { node } = createFormComponent(formProps);
-
-        // live validate does not run on initial render anymore
-        expect(node.querySelectorAll('.rjsf-field-error')).toHaveLength(0);
-
-        await user.clear(node.querySelector<HTMLInputElement>('input')!);
-        await user.type(node.querySelector<HTMLInputElement>('input')!, 'shorts');
-
-        expect(node.querySelectorAll('.rjsf-field-error')).toHaveLength(1);
-        expect(node.querySelector('.rjsf-field-string .error-detail')).toHaveTextContent(
-          'must NOT have fewer than 8 characters',
-        );
-      });
-    });
-
-    describe('root level with multiple errors, live validation', () => {
-      const formProps: Omit<FormProps, 'validator'> = {
-        liveValidate: true,
-        schema: {
-          type: 'string',
-          minLength: 8,
-          pattern: 'd+',
-        },
-        formData: 'short',
-      };
-
-      it('should reflect the contextualized error in state', async () => {
-        const { node, onError } = createFormComponent(formProps);
-        await submitForm(node, user);
-        expect(onError).toHaveBeenLastCalledWith([
-          {
-            message: 'must NOT have fewer than 8 characters',
-            name: 'minLength',
-            params: { limit: 8 },
-            property: '',
-            schemaPath: '#/minLength',
-            stack: 'must NOT have fewer than 8 characters',
-            title: '',
-          },
-          {
-            message: 'must match pattern "d+"',
-            name: 'pattern',
-            params: { pattern: 'd+' },
-            property: '',
-            schemaPath: '#/pattern',
-            stack: 'must match pattern "d+"',
-            title: '',
-          },
-        ]);
-      });
-
-      it('should denote the error in the field', async () => {
-        const { node } = createFormComponent(formProps);
-
-        // live validate does not run on initial render anymore
-        expect(node.querySelectorAll('.rjsf-field-error')).toHaveLength(0);
-
-        await user.clear(node.querySelector<HTMLInputElement>('input')!);
-        await user.type(node.querySelector<HTMLInputElement>('input')!, 'shorts');
-        const liNodes = node.querySelectorAll('.rjsf-field-string .error-detail li');
-        const errors = [].map.call(liNodes, (li: Element) => li.textContent);
-
-        expect(errors).toEqual(['must NOT have fewer than 8 characters', 'must match pattern "d+"']);
-      });
-    });
-
-    describe('nested field level, live validation', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          level1: {
-            type: 'object',
-            properties: {
-              level2: {
-                type: 'string',
-                minLength: 8,
-              },
-            },
-          },
-        },
-      };
-
-      const formProps: Omit<FormProps, 'validator'> = {
-        schema,
-        liveValidate: true,
-        formData: {
-          level1: {
-            level2: 'short',
-          },
-        },
-      };
-
-      it('should reflect the contextualized error in state', async () => {
-        const { node, onError } = createFormComponent(formProps);
-
-        await submitForm(node, user);
-        expect(onError).toHaveBeenLastCalledWith([
-          {
-            message: 'must NOT have fewer than 8 characters',
-            name: 'minLength',
-            params: { limit: 8 },
-            property: '.level1.level2',
-            schemaPath: '#/properties/level1/properties/level2/minLength',
-            stack: '.level1.level2 must NOT have fewer than 8 characters',
-            title: '',
-          },
-        ]);
-      });
-
-      it('should denote the error in the field', async () => {
-        const { node } = createFormComponent(formProps);
-        // live validate does not run on initial render anymore
-        expect(node.querySelectorAll('.rjsf-field-error')).toHaveLength(0);
-
-        await user.clear(node.querySelector<HTMLInputElement>('input')!);
-        await user.type(node.querySelector<HTMLInputElement>('input')!, 'shorts');
-        const errorDetail = node.querySelector('.rjsf-field-object .rjsf-field-string .error-detail');
-
-        expect(node.querySelectorAll('.rjsf-field-error')).toHaveLength(1);
-        expect(errorDetail).toHaveTextContent('must NOT have fewer than 8 characters');
-      });
-    });
-
-    describe('array indices, live validation', () => {
-      const schema: RJSFSchema = {
-        type: 'array',
-        items: {
-          type: 'string',
-          minLength: 4,
-        },
-      };
-
-      const formProps: Omit<FormProps, 'validator'> = {
-        schema,
-        liveValidate: true,
-        formData: ['good', 'ba', 'good'],
-      };
-
-      it('should contextualize the error for array indices', async () => {
-        const { node, onError } = createFormComponent(formProps);
-
-        await submitForm(node, user);
-        expect(onError).toHaveBeenLastCalledWith([
-          {
-            message: 'must NOT have fewer than 4 characters',
-            name: 'minLength',
-            params: { limit: 4 },
-            property: '.1',
-            schemaPath: '#/items/minLength',
-            stack: '.1 must NOT have fewer than 4 characters',
-            title: '',
-          },
-        ]);
-      });
-
-      it('should denote the error in the item field in error', async () => {
-        const { node } = createFormComponent(formProps);
-        const fieldNodes = node.querySelectorAll('.rjsf-field-string');
-
-        // live validate does not run on initial render anymore
-        expect(fieldNodes[1].classList.contains('rjsf-field-error')).toBe(false);
-
-        // Change the End field to a larger value than Start field to remove customValidate raised errors.
-        await user.clear(fieldNodes[1].querySelector('input')!);
-        await user.type(fieldNodes[1].querySelector('input')!, 'bad');
-
-        const liNodes = fieldNodes[1].querySelectorAll('.rjsf-field-string .error-detail li');
-        const errors = [].map.call(liNodes, (li: Element) => li.textContent);
-
-        expect(errors).toEqual(['must NOT have fewer than 4 characters']);
-      });
-
-      it('should not denote errors on non impacted fields', () => {
-        const { node } = createFormComponent(formProps);
-        const fieldNodes = node.querySelectorAll('.rjsf-field-string');
-
-        expect(fieldNodes[0].classList.contains('rjsf-field-error')).toBe(false);
-        expect(fieldNodes[2].classList.contains('rjsf-field-error')).toBe(false);
-      });
-    });
-
-    describe('nested array indices, live validation', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          level1: {
-            type: 'array',
-            items: {
-              type: 'string',
-              minLength: 4,
-            },
-          },
-        },
-      };
-
-      const formProps: Omit<FormProps, 'validator'> = { schema, liveValidate: true };
-
-      it('should contextualize the error for nested array indices', async () => {
-        const { node, onError } = createFormComponent({
-          ...formProps,
-          formData: {
-            level1: ['good', 'bad', 'good', 'bad'],
-          },
-        });
-        await submitForm(node, user);
-        expect(onError).toHaveBeenLastCalledWith([
-          {
-            message: 'must NOT have fewer than 4 characters',
-            name: 'minLength',
-            params: { limit: 4 },
-            property: '.level1.1',
-            schemaPath: '#/properties/level1/items/minLength',
-            stack: '.level1.1 must NOT have fewer than 4 characters',
-            title: '',
-          },
-          {
-            message: 'must NOT have fewer than 4 characters',
-            name: 'minLength',
-            params: { limit: 4 },
-            property: '.level1.3',
-            schemaPath: '#/properties/level1/items/minLength',
-            stack: '.level1.3 must NOT have fewer than 4 characters',
-            title: '',
-          },
-        ]);
-      });
-
-      it('should denote the error in the nested item field in error', async () => {
-        const { node } = createFormComponent({
-          ...formProps,
-          formData: {
-            level1: ['good', 'ba', 'good'],
-          },
-        });
-
-        const fields = node.querySelectorAll('.rjsf-field-string');
-        let liNodes = node.querySelectorAll('.rjsf-field-string .error-detail li');
-        let errors = [].map.call(liNodes, (li: Element) => li.textContent);
-
-        // live validate does not run on initial render anymore
-        expect(errors).toEqual([]);
-
-        await user.clear(fields[1].querySelector('input')!);
-        await user.type(fields[1].querySelector('input')!, 'bad');
-
-        liNodes = node.querySelectorAll('.rjsf-field-string .error-detail li');
-        errors = [].map.call(liNodes, (li: Element) => li.textContent);
-        expect(errors).toEqual(['must NOT have fewer than 4 characters']);
-      });
-    });
-
-    describe('nested arrays, live validation', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          outer: {
-            type: 'array',
-            items: {
-              type: 'array',
-              items: {
-                type: 'string',
-                minLength: 4,
-              },
-            },
-          },
-        },
-      };
-
-      const formData = {
-        outer: [
-          ['good', 'bad'],
-          ['bad', 'good'],
-        ],
-      };
-
-      const formProps: Omit<FormProps, 'validator'> = { schema, formData, liveValidate: true };
-
-      it('should contextualize the error for nested array indices, focusing on first error', async () => {
-        const { node, onError } = createFormComponent({
-          ...formProps,
-          focusOnFirstError: true,
-        });
-
-        const focusSpy = vi.fn();
-        const input = node.querySelector<HTMLInputElement>('input[id=root_outer_0_1]')!;
-        // Since programmatically triggering focus does not call onFocus, change the focus method to a spy
-        Object.defineProperty(input, 'focus', {
-          configurable: true,
-          value: focusSpy,
-        });
-
-        await submitForm(node, user);
-        expect(onError).toHaveBeenLastCalledWith([
-          {
-            message: 'must NOT have fewer than 4 characters',
-            name: 'minLength',
-            params: { limit: 4 },
-            property: '.outer.0.1',
-            schemaPath: '#/properties/outer/items/items/minLength',
-            stack: '.outer.0.1 must NOT have fewer than 4 characters',
-            title: '',
-          },
-          {
-            message: 'must NOT have fewer than 4 characters',
-            name: 'minLength',
-            params: { limit: 4 },
-            property: '.outer.1.0',
-            schemaPath: '#/properties/outer/items/items/minLength',
-            stack: '.outer.1.0 must NOT have fewer than 4 characters',
-            title: '',
-          },
-        ]);
-        expect(focusSpy).toHaveBeenCalledTimes(1);
-      });
-
-      it('should denote the error in the nested item field in error', async () => {
-        const { node } = createFormComponent(formProps);
-
-        const fields = node.querySelectorAll('.rjsf-field-string');
-        let errors = [].map.call(fields, (field: Element) => {
-          const li = field.querySelector('.error-detail li');
-          return li && li.textContent;
-        });
-
-        // live validate does not run on initial render anymore
-        expect(errors).toEqual([null, null, null, null]);
-
-        await user.clear(fields[0].querySelector('input')!);
-        await user.type(fields[0].querySelector('input')!, 'bad');
-        errors = [].map.call(fields, (field: Element) => {
-          const li = field.querySelector('.error-detail li');
-          return li && li.textContent;
-        });
-
-        expect(errors).toEqual([
-          'must NOT have fewer than 4 characters',
-          'must NOT have fewer than 4 characters',
-          'must NOT have fewer than 4 characters',
-          null,
-        ]);
-      });
-    });
-
-    describe('array nested items, live validation', () => {
-      const schema: RJSFSchema = {
-        type: 'array',
-        items: {
-          type: 'object',
-          properties: {
-            foo: {
-              type: 'string',
-              minLength: 4,
-            },
-          },
-        },
-      };
-
-      const formProps: Omit<FormProps, 'validator'> = {
-        schema,
-        liveValidate: true,
-        formData: [{ foo: 'good' }, { foo: 'ba' }, { foo: 'good' }],
-      };
-
-      it('should contextualize the error for array nested items', async () => {
-        const { node, onError } = createFormComponent(formProps);
-
-        await submitForm(node, user);
-        expect(onError).toHaveBeenLastCalledWith([
-          {
-            message: 'must NOT have fewer than 4 characters',
-            name: 'minLength',
-            params: { limit: 4 },
-            property: '.1.foo',
-            schemaPath: '#/items/properties/foo/minLength',
-            stack: '.1.foo must NOT have fewer than 4 characters',
-            title: '',
-          },
-        ]);
-      });
-
-      it('should denote the error in the array nested item', async () => {
-        const { node } = createFormComponent(formProps);
-        const fieldNodes = node.querySelectorAll('.rjsf-field-string');
-
-        // Initial render no longer does live validation
-        expect(fieldNodes[1].classList.contains('rjsf-field-error')).toBe(false);
-        // Change the End field to a larger value than Start field to remove customValidate raised errors.
-        await user.clear(fieldNodes[1].querySelector('input')!);
-        await user.type(fieldNodes[1].querySelector('input')!, 'bad');
-
-        const liNodes = fieldNodes[1].querySelectorAll('.error-detail li');
-        const errors = [].map.call(liNodes, (li: Element) => li.textContent);
-        expect(fieldNodes[1].classList.contains('rjsf-field-error')).toEqual(true);
-        expect(errors).toEqual(['must NOT have fewer than 4 characters']);
-      });
-    });
-
-    describe('schema dependencies, live validation', () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          branch: {
-            type: 'number',
-            enum: [1, 2, 3],
-            default: 1,
-          },
-        },
-        required: ['branch'],
-        dependencies: {
-          branch: {
-            oneOf: [
-              {
-                properties: {
-                  branch: {
-                    enum: [1],
-                  },
-                  field1: {
-                    type: 'number',
-                  },
-                },
-                required: ['field1'],
-              },
-              {
-                properties: {
-                  branch: {
-                    enum: [2],
-                  },
-                  field1: {
-                    type: 'number',
-                  },
-                  field2: {
-                    type: 'number',
-                  },
-                },
-                required: ['field1', 'field2'],
-              },
-            ],
-          },
-        },
-      };
-
-      it('should only show error for property in selected branch', async () => {
-        const { node, onChange } = createFormComponent({
-          schema,
-          liveValidate: true,
-        });
-
-        const input = node.querySelector<HTMLInputElement>('input[type=number]');
-        expect(input).toBeInTheDocument();
-        await user.type(input!, '0');
-        await user.clear(input!);
-
-        expect(onChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            errorSchema: { field1: { __errors: ["must have required property 'field1'"] } },
-          }),
-          'root_field1',
-        );
-      });
-
-      it('should only show errors for properties in selected branch', async () => {
-        const { node, onChange } = createFormComponent({
-          ref: createRef(),
-          schema,
-          liveValidate: true,
-          formData: { branch: 2 },
-        });
-
-        await user.type(node.querySelectorAll<HTMLInputElement>('input[type=number]')[0], '0');
-
-        expect(onChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            errorSchema: {
-              field2: {
-                __errors: ["must have required property 'field2'"],
-              },
-            },
-          }),
-          'root_field1',
-        );
-      });
-
-      it('should not show any errors when branch is empty', async () => {
-        const consoleWarnSpy = vi.spyOn(console, 'warn').mockImplementation(noop);
-        const { node, onChange } = createFormComponent({
-          ref: createRef(),
-          schema,
-          liveValidate: true,
-          formData: { branch: 3 },
-        });
-
-        await user.selectOptions(node.querySelector('select')!, '3'); // Select by label/text since selectOptions matches text before value attribute
-
-        expect(onChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            errorSchema: {},
-          }),
-          'root_branch',
-        );
-        expect(consoleWarnSpy).toHaveBeenCalledWith(
-          "ignoring oneOf in dependencies because there isn't exactly one subschema that is valid",
-        );
-        consoleWarnSpy.mockRestore();
-      });
-
-      it('should sanitize stale enum data and persist the retrieved dependency schema', async () => {
-        const formRef = createRef<Form>();
-        const dependentEnumSchema: RJSFSchema = {
-          type: 'object',
-          properties: {
-            animal: {
-              type: 'string',
-              enum: ['Cat', 'Fish'],
-            },
-          },
-          dependencies: {
-            animal: {
-              oneOf: [
-                {
-                  properties: {
-                    animal: {
-                      enum: ['Cat'],
-                    },
-                    food: {
-                      type: 'string',
-                      enum: ['meat'],
-                    },
-                  },
-                },
-                {
-                  properties: {
-                    animal: {
-                      enum: ['Fish'],
-                    },
-                    food: {
-                      type: 'string',
-                      enum: ['worms'],
-                    },
-                    water: {
-                      type: 'string',
-                      enum: ['lake'],
-                    },
-                  },
-                },
-              ],
-            },
-          },
-        };
-        const { node, onChange } = createFormComponent({
-          ref: formRef,
-          schema: dependentEnumSchema,
-          formData: { animal: 'Fish', food: 'worms', water: 'lake' },
-        });
-
-        await user.selectOptions(node.querySelector<HTMLSelectElement>('#root_animal')!, '0');
-
-        expectToHaveBeenCalledWithFormData(onChange, { animal: 'Cat', food: 'meat', water: undefined }, 'root_animal');
-        const { retrievedSchema } = formRef.current!.state;
-        expect(retrievedSchema.properties).toEqual(
-          expect.objectContaining({
-            food: expect.objectContaining({ enum: ['meat'] }),
-          }),
-        );
-        expect(retrievedSchema.properties).not.toHaveProperty('water');
-
-        await user.selectOptions(node.querySelector<HTMLSelectElement>('#root_food')!, '0');
-        expect(formRef.current!.state.retrievedSchema.properties).not.toHaveProperty('water');
-      });
-    });
-
-    describe('customValidate errors, live validation', () => {
-      it('customValidate should raise an error when End is larger than Start field.', async () => {
-        const schema: RJSFSchema = {
-          required: ['Start', 'End'],
-          properties: {
-            Start: {
-              type: 'number',
-            },
-            End: {
-              type: 'number',
-            },
-          },
-          type: 'object',
-        };
-
-        // customValidate method to raise an error when Start is larger than End field.
-        const customValidate = (formData: any, errors: FormValidation) => {
-          if (formData.Start > formData.End) {
-            errors.Start?.addError('Validate error: Test should be LE than End');
-          }
-          return errors;
-        };
-
-        const { node, onChange } = createFormComponent({
-          schema,
-          liveValidate: true,
-          formData: { Start: 2, End: 0 },
-          customValidate,
-        });
-
-        // live validate does not run on initial render anymore
-        expect(node.querySelectorAll('#root_Start__error')).toHaveLength(0);
-
-        // Change the End field to a larger value than Start field to remove customValidate raised errors.
-        const endInput = node.querySelector<HTMLInputElement>('#root_End')!;
-        await user.clear(endInput);
-        await user.type(endInput, '1');
-        expect(node.querySelectorAll('#root_Start__error')).toHaveLength(1);
-        let errorMessageContent = node.querySelector('#root_Start__error .text-danger');
-        expect(errorMessageContent).toHaveTextContent('Validate error: Test should be LE than End');
-
-        // Change the End field to a larger value than Start field to remove customValidate raised errors.
-        await user.clear(endInput);
-        await user.type(endInput, '3');
-
-        expect(node.querySelectorAll('#root_Start__error')).toHaveLength(0);
-        expect(onChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            errors: [],
-          }),
-          'root_End',
-        );
-
-        // Change the End field to a lesser value than Start field to raise customValidate errors.
-        await user.clear(endInput);
-        await user.type(endInput, '0');
-
-        expect(node.querySelectorAll('#root_Start__error')).toHaveLength(1);
-        errorMessageContent = node.querySelector('#root_Start__error .text-danger');
-        expect(errorMessageContent).toHaveTextContent('Validate error: Test should be LE than End');
-        expect(onChange).toHaveBeenLastCalledWith(
-          expect.objectContaining({
-            errorSchema: expect.objectContaining({
-              Start: {
-                __errors: ['Validate error: Test should be LE than End'],
-              },
-            }),
-          }),
-          'root_End',
-        );
-      });
-    });
-  });
-
-  describe('Schema and formData updates', () => {
-    // https://github.com/rjsf-team/react-jsonschema-form/issues/231
-    const schema: RJSFSchema = {
-      type: 'object',
-      properties: {
-        foo: { type: 'string' },
-        bar: { type: 'string' },
-      },
-    };
-
-    it('should replace state when props remove formData keys', async () => {
-      const formData = { foo: 'foo', bar: 'bar' };
-      const { node, onChange, rerender } = createFormComponent({
-        ref: createRef(),
-        schema,
-        formData,
-      });
-
-      rerender({
-        onChange,
-        schema: {
-          type: 'object',
-          properties: {
-            bar: { type: 'string' },
-          },
-        },
-        formData: { bar: 'bar' },
-      });
-
-      await user.clear(node.querySelector('#root_bar')!);
-      await user.type(node.querySelector('#root_bar')!, 'baz');
-
-      expectToHaveBeenCalledWithFormData(onChange, { bar: 'baz' }, 'root_bar');
-    });
-
-    it('should replace state when props change formData keys', async () => {
-      const formData = { foo: 'foo', bar: 'bar' };
-      const { node, onChange, rerender } = createFormComponent({
-        ref: createRef(),
-        schema,
-        formData,
-      });
-
-      rerender({
-        onChange,
-        schema: {
-          type: 'object',
-          properties: {
-            foo: { type: 'string' },
-            baz: { type: 'string' },
-          },
-        },
-        formData: { foo: 'foo', baz: 'bar' },
-      });
-
-      await user.clear(node.querySelector('#root_baz')!);
-      await user.type(node.querySelector('#root_baz')!, 'baz');
-
-      expectToHaveBeenCalledWithFormData(onChange, { foo: 'foo', baz: 'baz' }, 'root_baz');
-    });
-  });
-  describe('Form disable prop', () => {
-    const schema: RJSFSchema = {
-      type: 'object',
-      properties: {
-        foo: { type: 'string' },
-        bar: { type: 'string' },
-      },
-    };
-    const formData = { foo: 'foo', bar: 'bar' };
-
-    it('should enable all items', () => {
-      const { node } = createFormComponent({ schema, formData });
-
-      expect(node.querySelectorAll('input:disabled')).toHaveLength(0);
-    });
-
-    it('should disable all items', () => {
-      const { node } = createFormComponent({
-        schema,
-        formData,
-        disabled: true,
-      });
-
-      expect(node.querySelectorAll('input:disabled')).toHaveLength(2);
-    });
-
-    it('should disable the submit button', () => {
-      const { node } = createFormComponent({
-        schema,
-        formData,
-        disabled: true,
-      });
-
-      expect(node.querySelector("button[type='submit']")).toBeInTheDocument();
-      expect(node.querySelector("button[type='submit']:disabled")).toBeInTheDocument();
-    });
-
-    it('disabling the submit button via ui:schema - ui:submitButtonOptions props is still possible', () => {
-      const { node } = createFormComponent({
-        schema,
-        formData,
-        uiSchema: { 'ui:submitButtonOptions': { props: { disabled: true } } },
-      });
-
-      expect(node.querySelector("button[type='submit']")).toBeInTheDocument();
-      expect(node.querySelector("button[type='submit']:disabled")).toBeInTheDocument();
-    });
-
-    it('disabling the submit button via ui:schema - ui:options, submitButtonOptions props is still possible', () => {
-      const { node } = createFormComponent({
-        schema,
-        formData,
-        uiSchema: { 'ui:options': { submitButtonOptions: { submitText: 'hello', props: { disabled: true } } } },
-      });
-
-      expect(node.querySelector("button[type='submit']")).toBeInTheDocument();
-      expect(node.querySelector("button[type='submit']")).toHaveTextContent('hello');
-      expect(node.querySelector("button[type='submit']:disabled")).toBeInTheDocument();
-    });
-
-    it('if both ui:submitButtonProps and the main form disabled props are provided, and either of them are true, the button will be disabled', () => {
-      const { node } = createFormComponent({
-        schema,
-        formData,
-        uiSchema: { 'ui:submitButtonOptions': { props: { disabled: false } } },
-        disabled: true,
-      });
-
-      expect(node.querySelector("button[type='submit']")).toBeInTheDocument();
-      expect(node.querySelector("button[type='submit']:disabled")).toBeInTheDocument();
-    });
-
-    it('if both ui:submitButtonProps and the main form disabled props are provided, but false, then submit button will not be disabled', () => {
-      const { node } = createFormComponent({
-        schema,
-        formData,
-        uiSchema: { 'ui:submitButtonOptions': { props: { disabled: false } } },
-        disabled: false,
-      });
-
-      expect(node.querySelector("button[type='submit']")).toBeInTheDocument();
-      expect(node.querySelector("button[type='submit']:disabled")).not.toBeInTheDocument();
-    });
-  });
-
-  describe('Form readonly prop', () => {
-    const schema: RJSFSchema = {
-      type: 'object',
-      properties: {
-        foo: { type: 'string' },
-        bar: { type: 'object', properties: { baz: { type: 'string' } } },
-      },
-    };
-    const formData = { foo: 'foo', bar: { baz: 'baz' } };
-
-    it('should not have any readonly items', () => {
-      const { node } = createFormComponent({ schema, formData });
-
-      expect(node.querySelectorAll('input:read-only')).toHaveLength(0);
-    });
-
-    it('should readonly all items', () => {
-      const { node } = createFormComponent({
-        schema,
-        formData,
-        readonly: true,
-      });
-
-      expect(node.querySelectorAll('input:read-only')).toHaveLength(2);
-    });
-  });
-
-  describe('Attributes', () => {
-    const formProps: Omit<FormProps, 'validator'> = {
-      schema: {},
-      id: 'test-form',
-      className: 'test-class other-class',
-      name: 'testName',
-      method: 'post',
-      target: '_blank',
-      action: '/users/list',
-      autoComplete: 'off',
-      enctype: 'multipart/form-data',
-      acceptCharset: 'ISO-8859-1',
-      noHtml5Validate: true,
-    };
-
-    let node: Element;
-
-    beforeEach(() => {
-      node = createFormComponent(formProps).node;
-    });
-
-    it('should set attr id of form', () => {
-      expect(node.getAttribute('id')).toEqual(formProps.id);
-    });
-
-    it('should set attr class of form', () => {
-      expect(node.getAttribute('class')).toEqual(formProps.className);
-    });
-
-    it('should set attr name of form', () => {
-      expect(node).toHaveAttribute('name', formProps.name);
-    });
-
-    it('should set attr method of form', () => {
-      expect(node.getAttribute('method')).toEqual(formProps.method);
-    });
-
-    it('should set attr target of form', () => {
-      expect(node.getAttribute('target')).toEqual(formProps.target);
-    });
-
-    it('should set attr action of form', () => {
-      expect(node.getAttribute('action')).toEqual(formProps.action);
-    });
-
-    it('should set attr enctype of form', () => {
-      expect(node.getAttribute('enctype')).toEqual(formProps.enctype);
-    });
-
-    it('should set attr acceptCharset of form', () => {
-      expect(node.getAttribute('accept-charset')).toEqual(formProps.acceptCharset);
-    });
-
-    it('should set attr novalidate of form', () => {
-      expect(node.getAttribute('novalidate')).not.toBeNull();
-    });
-  });
-
-  describe('Custom format updates, live validation', () => {
-    it('Should update custom formats when customFormats is changed', async () => {
-      const consoleWarnSpy = vi.spyOn(console, 'warn').mockImplementation(noop);
-      const formProps: Omit<FormProps, 'validator'> = {
-        ref: createRef(),
-        liveValidate: true,
-        formData: {
-          areaCode: '123455',
-        },
-        schema: {
-          type: 'object',
-          properties: {
-            areaCode: {
-              type: 'string',
-              format: 'area-code',
-            },
-          },
-        },
-        uiSchema: {
-          areaCode: {
-            'ui:widget': 'area-code',
-          },
-        },
-        widgets: {
-          'area-code': () => <div id='custom' />,
-        },
-      };
-
-      const customValidator = customizeValidator({
-        customFormats: {
-          'area-code': /^\d{3}$/,
-        },
-      });
-
-      const { node, onError, rerender } = createFormComponent(formProps);
-
-      await submitForm(node, user);
-      expect(onError).not.toHaveBeenCalled();
-
-      rerender(
-        {
-          ...formProps,
-          onError,
-        },
-        customValidator,
-      );
-
-      await submitForm(node, user);
-      expect(onError).toHaveBeenLastCalledWith([
-        {
-          message: 'must match format "area-code"',
-          name: 'format',
-          params: { format: 'area-code' },
-          property: '.areaCode',
-          schemaPath: '#/properties/areaCode/format',
-          stack: '.areaCode must match format "area-code"',
-          title: '',
-        },
-      ]);
-      expect(consoleWarnSpy).toHaveBeenCalledWith(
-        expect.stringContaining('unknown format "area-code" ignored in schema at path "#/properties/areaCode"'),
-      );
-      consoleWarnSpy.mockRestore();
-    });
-  });
-
-  describe('Meta schema updates', () => {
-    it('Should update allowed meta schemas when additionalMetaSchemas is changed', async () => {
-      const formProps: Omit<FormProps, 'validator'> = {
-        ref: createRef(),
-        schema: {
-          $schema: 'http://json-schema.org/draft-06/schema#',
-          type: 'string',
-          minLength: 8,
-          pattern: 'd+',
-        },
-        formData: 'short',
-      };
-
-      const { node, onError, rerender } = createFormComponent(formProps);
-
-      await submitForm(node, user);
-      expect(onError).toHaveBeenLastCalledWith([
-        {
-          stack: 'no schema with key or ref "http://json-schema.org/draft-06/schema#"',
-        },
-      ]);
-
-      const customValidator = customizeValidator({
-        additionalMetaSchemas: [draft06],
-      });
-
-      rerender(
-        {
-          ...formProps,
-          onError,
-        },
-        customValidator,
-      );
-
-      await submitForm(node, user);
-      expect(onError).toHaveBeenLastCalledWith([
-        {
-          message: 'must NOT have fewer than 8 characters',
-          name: 'minLength',
-          params: { limit: 8 },
-          property: '',
-          schemaPath: '#/minLength',
-          stack: 'must NOT have fewer than 8 characters',
-          title: '',
-        },
-        {
-          message: 'must match pattern "d+"',
-          name: 'pattern',
-          params: { pattern: 'd+' },
-          property: '',
-          schemaPath: '#/pattern',
-          stack: 'must match pattern "d+"',
-          title: '',
-        },
-      ]);
-    });
-  });
-
-  describe('Changing the tagName', () => {
-    it('should render the component using the custom tag name', () => {
-      const tagName = 'span';
-      const { node } = createFormComponent({ schema: {}, tagName });
-      expect(node.tagName).toEqual(tagName.toUpperCase());
-    });
-
-    it('should render the component using a ComponentType', () => {
-      const Component = (props: any) => <div {...props} id='test' />;
-      const { node } = createFormComponent({ schema: {}, tagName: Component });
-      expect(node.id).toEqual('test');
-      // React deduplicates this warning per component name — only fires on the first test iteration
-      if (renderErrorSuppression.consoleSpy.mock.calls.length > 0) {
-        expect(renderErrorSuppression.consoleSpy).toHaveBeenCalledWith(
-          expect.stringContaining('Function components cannot be given refs'),
-          expect.any(String),
-          expect.any(String),
-        );
-      }
-    });
-  });
-
-  describe('Nested forms', () => {
-    it('should call provided submit handler with form state', async () => {
-      const innerOnSubmit = vi.fn();
-      const outerOnSubmit = vi.fn();
-      const innerRef = createRef<HTMLDivElement>();
-
-      const ArrayTemplateWithForm = (_props: FormProps) => {
-        const innerFormProps = {
-          validator,
-          schema: {},
-          onSubmit: innerOnSubmit,
-        };
-
-        return createPortal(
-          <div className='array' ref={innerRef}>
-            <Form {...innerFormProps}>
-              <button className='array-form-submit' type='submit'>
-                Submit
-              </button>
-            </Form>
-          </div>,
-          document.body,
-        );
-      };
-
-      createFormComponent({
-        schema: {
-          type: 'array',
-          title: 'my list',
-          description: 'my description',
-          items: { type: 'string' },
-        },
-        formData: ['foo', 'bar'],
-        templates: { ArrayFieldTemplate: ArrayTemplateWithForm },
-        onSubmit: outerOnSubmit,
-      });
-      expect(innerRef).toBeDefined();
-      expect(innerRef.current).not.toBeNull();
-      const arrayForm = innerRef.current!.querySelector('form')!;
-      const arraySubmit = arrayForm.querySelector<HTMLButtonElement>('.array-form-submit')!;
-
-      await user.click(arraySubmit);
-
-      expect(innerOnSubmit).toHaveBeenCalledTimes(1);
-      expect(outerOnSubmit).not.toHaveBeenCalled();
-    });
-  });
-
-  describe('Dependencies', () => {
-    it('should not give a validation error by duplicating enum values in dependencies', async () => {
-      const schema: RJSFSchema = {
-        title: 'A registration form',
-        description: 'A simple form example.',
-        type: 'object',
-        properties: {
-          type1: {
-            type: 'string',
-            title: 'Type 1',
-            enum: ['FOO', 'BAR', 'BAZ'],
-          },
-          type2: {
-            type: 'string',
-            title: 'Type 2',
-            enum: ['GREEN', 'BLUE', 'RED'],
-          },
-        },
-        dependencies: {
-          type1: {
-            properties: {
-              type1: {
-                enum: ['FOO'],
-              },
-              type2: {
-                enum: ['GREEN'],
-              },
-            },
-          },
-        },
-      };
-      const formData = {
-        type1: 'FOO',
-      };
-      const { node, onError } = createFormComponent({ schema, formData });
-      await submitForm(node, user);
-      expect(onError).not.toHaveBeenCalled();
-    });
-    it('should show dependency defaults for uncontrolled components', async () => {
-      const schema: RJSFSchema = {
-        type: 'object',
-        properties: {
-          firstName: { type: 'string' },
-        },
-        dependencies: {
-          firstName: {
-            properties: {
-              lastName: { type: 'string', default: 'Norris' },
-            },
-          },
-        },
-      };
-      const { node } = createFormComponent({ schema });
-
-      await user.type(node.querySelector('#root_firstName')!, 'Chuck');
-      expect(node.querySelector('#root_lastName')).toHaveValue('Norris');
-    });
-  });
-});
+setupConsoleErrorSuppression();
 
 describe('Error paths that collide with prototype keys', () => {
   it('renders the inline error for a field literally named "constructor"', async () => {
