+++ packages/core/test/FormCommon.handlers.test.tsx
@@ -1,0 +1,730 @@
+import { createRef, useEffect } from 'react';
+import type { RJSFSchema, UiSchema, WidgetProps } from '@rjsf/utils';
+import { getTemplate, getUiOptions } from '@rjsf/utils';
+import { waitFor } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+
+import type { FormProps, IChangeEvent } from '../src';
+import { expectToHaveBeenCalledWithFormData, submitForm, describeRepeated } from './testUtils';
+
+const user = userEvent.setup();
+
+describeRepeated('Form common: event handlers', (createFormComponent) => {
+  describe('Submit handler', () => {
+    it('should call provided submit handler with form state', async () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        properties: {
+          foo: { type: 'string' },
+        },
+      };
+      const formData = {
+        foo: 'bar',
+      };
+      const { node, onSubmit } = createFormComponent({
+        ref: createRef(),
+        schema,
+        formData,
+      });
+
+      await submitForm(node, user);
+      expectToHaveBeenCalledWithFormData(onSubmit, formData, true);
+    });
+
+    it('should not call provided submit handler on validation errors', async () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        properties: {
+          foo: {
+            type: 'string',
+            minLength: 1,
+          },
+        },
+      };
+      const formData = {
+        foo: '',
+      };
+      const { node, onSubmit, onError } = createFormComponent({
+        ref: createRef(),
+        schema,
+        formData,
+      });
+
+      await submitForm(node, user);
+
+      expect(onSubmit).not.toHaveBeenCalled();
+      expect(onError).toHaveBeenCalled();
+    });
+  });
+
+  describe('Change handler', () => {
+    it('should call provided change handler on form state change with schema and uiSchema', async () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        properties: {
+          foo: {
+            type: 'string',
+          },
+        },
+      };
+      const uiSchema: UiSchema = {
+        foo: { 'ui:field': 'textarea' },
+      };
+
+      const formData = {
+        foo: '',
+      };
+      const { node, onChange } = createFormComponent({
+        ref: createRef(),
+        schema,
+        uiSchema,
+        formData,
+      });
+
+      await user.type(node.querySelector('[type=text]')!, 'new');
+
+      expectToHaveBeenCalledWithFormData(onChange, { foo: 'new' }, 'root_foo');
+    });
+    it('should call last provided change handler', async () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        properties: {
+          foo: {
+            type: 'string',
+            default: 'bar',
+          },
+        },
+      };
+
+      const secondOnChange = vi.fn();
+
+      const { onChange, rerender } = createFormComponent({ ref: createRef(), schema, formData: { foo: 'bar1' } });
+
+      rerender({ schema, formData: {}, onChange });
+
+      expect(onChange).toHaveBeenCalledTimes(1);
+
+      rerender({ schema, formData: { foo: 'bar2' } });
+
+      rerender({ schema, formData: {}, onChange: secondOnChange });
+
+      expect(onChange).toHaveBeenCalledTimes(1);
+      expect(secondOnChange).toHaveBeenCalledTimes(1);
+    });
+    it('should call change handler with proper data after two near simultaneous changes', async () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        properties: {
+          foo: {
+            type: 'string',
+            default: 'bar',
+          },
+          baz: {
+            type: 'string',
+            default: 'blah',
+          },
+        },
+      };
+      function FooWidget(props: WidgetProps) {
+        const { value, id, onChange, uiSchema, registry } = props;
+        const uiOptions = getUiOptions(uiSchema);
+        const BaseInputTemplate = getTemplate('BaseInputTemplate', registry, uiOptions);
+        useEffect(() => {
+          if (value === 'bar') {
+            onChange('bar2', undefined, id);
+          }
+        }, [value, onChange, id]);
+        return <BaseInputTemplate {...props} />;
+      }
+      function BazWidget(props: WidgetProps) {
+        const { value, id, onChange, uiSchema, registry } = props;
+        const uiOptions = getUiOptions(uiSchema);
+        const BaseInputTemplate = getTemplate('BaseInputTemplate', registry, uiOptions);
+        useEffect(() => {
+          if (value === 'blah') {
+            onChange('blah2', undefined, id);
+          }
+        }, [value, onChange, id]);
+        return <BaseInputTemplate {...props} />;
+      }
+      const uiSchema: UiSchema = {
+        foo: {
+          'ui:widget': FooWidget,
+        },
+        baz: {
+          'ui:widget': BazWidget,
+        },
+      };
+
+      let formData = {};
+      const ids: (string | undefined)[] = [];
+      const onChange: FormProps['onChange'] = (data, id) => {
+        const { formData: fd } = data;
+        formData = { ...formData, ...fd };
+        ids.push(id);
+      };
+      createFormComponent({
+        schema,
+        formData,
+        onChange,
+        uiSchema,
+      });
+
+      await waitFor(() => {
+        expect(ids).toHaveLength(3);
+      });
+
+      expect(formData).toEqual({ foo: 'bar2', baz: 'blah2' });
+      // There will be 3 ids, undefined for the setting of the defaults and then the two updated components
+      expect(ids).toEqual([undefined, 'root_foo', 'root_baz']);
+    });
+    it('should modify an allOf field when the defaults are set', async () => {
+      const schema: RJSFSchema = {
+        properties: {
+          all_of_field: {
+            allOf: [
+              {
+                properties: {
+                  first: {
+                    type: 'string',
+                  },
+                },
+              },
+              {
+                properties: {
+                  second: {
+                    type: 'string',
+                  },
+                },
+              },
+            ],
+            default: {
+              second: 'second!',
+            },
+          },
+        },
+        type: 'object',
+      };
+
+      const { node, onChange } = createFormComponent({
+        schema,
+      });
+
+      const secondInputID = '#root_all_of_field_second';
+      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'second!');
+
+      await user.clear(node.querySelector(secondInputID)!);
+      await user.type(node.querySelector(secondInputID)!, 'changed!');
+
+      expectToHaveBeenCalledWithFormData(
+        onChange,
+        {
+          all_of_field: {
+            second: 'changed!',
+          },
+        },
+        'root_all_of_field_second',
+      );
+
+      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'changed!');
+    });
+    it('should modify an oneOf field when the defaults are set', async () => {
+      const schema: RJSFSchema = {
+        properties: {
+          one_of_field: {
+            oneOf: [
+              {
+                properties: {
+                  first: {
+                    type: 'string',
+                  },
+                },
+              },
+              {
+                properties: {
+                  second: {
+                    type: 'string',
+                  },
+                },
+              },
+            ],
+            default: {
+              second: 'second!',
+            },
+          },
+        },
+        type: 'object',
+      };
+
+      const { node, onChange } = createFormComponent({
+        schema,
+      });
+
+      const secondInputID = '#root_one_of_field_second';
+      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'second!');
+
+      await user.clear(node.querySelector(secondInputID)!);
+      await user.type(node.querySelector(secondInputID)!, 'changed!');
+
+      expectToHaveBeenCalledWithFormData(
+        onChange,
+        {
+          one_of_field: {
+            second: 'changed!',
+          },
+        },
+        'root_one_of_field_second',
+      );
+
+      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'changed!');
+    });
+    it('should modify an anyOf field when the defaults are set', async () => {
+      const schema: RJSFSchema = {
+        properties: {
+          any_of_field: {
+            anyOf: [
+              {
+                properties: {
+                  first: {
+                    type: 'string',
+                  },
+                },
+              },
+              {
+                properties: {
+                  second: {
+                    type: 'string',
+                  },
+                },
+              },
+            ],
+            default: {
+              second: 'second!',
+            },
+          },
+        },
+        type: 'object',
+      };
+
+      const { node, onChange } = createFormComponent({
+        schema,
+      });
+
+      const secondInputID = '#root_any_of_field_second';
+      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'second!');
+
+      await user.clear(node.querySelector(secondInputID)!);
+      await user.type(node.querySelector(secondInputID)!, 'changed!');
+
+      expectToHaveBeenCalledWithFormData(
+        onChange,
+        {
+          any_of_field: {
+            second: 'changed!',
+          },
+        },
+        'root_any_of_field_second',
+      );
+
+      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'changed!');
+    });
+    it('should restore defaults when switching from null back to object option in oneOf', async () => {
+      // This test verifies that when switching from a null oneOf option back to an object option,
+      // the defaults are correctly restored. Without the fix, the form would show empty/undefined values.
+      const schema: RJSFSchema = {
+        type: 'object',
+        title: 'Configuration',
+        oneOf: [
+          {
+            title: 'Default Configuration',
+            type: 'object',
+            properties: {
+              types: { const: 'default', title: 'Types' },
+              content: { type: 'string', title: 'Content' },
+            },
+            required: ['types'],
+          },
+          {
+            title: 'Advanced Configuration',
+            type: 'object',
+            properties: {
+              types: { const: 'advanced', title: 'Types' },
+              content: { type: 'string', title: 'Content' },
+            },
+            required: ['types'],
+          },
+          { title: 'No Configuration', type: 'null' },
+        ],
+        default: { types: 'advanced', content: 'placeholder' },
+      };
+
+      const onChangeCalls: { event: IChangeEvent; id?: string }[] = [];
+
+      const { node } = createFormComponent({
+        schema,
+        onChange: (event: IChangeEvent, id?: string) => onChangeCalls.push({ event, id }),
+        experimental_defaultFormStateBehavior: { emptyObjectFields: 'populateAllDefaults' },
+      });
+
+      // Should start with "Advanced Configuration" (index 1) based on default
+      const oneOfSelect = node.querySelector<HTMLSelectElement>('#root__oneof_select');
+      expect(oneOfSelect).toBeInTheDocument();
+      expect(oneOfSelect!.value).toEqual('1');
+
+      // The content field should have the default value
+      let contentInput = node.querySelector<HTMLInputElement>('#root_content');
+      expect(contentInput).toBeInTheDocument();
+      expect(contentInput!.value).toEqual('placeholder');
+
+      // Switch to "No Configuration" (null option, index 2)
+      await user.selectOptions(oneOfSelect!, '2');
+
+      // Verify we're now on null option - content field should not exist
+      expect(node.querySelector<HTMLSelectElement>('#root__oneof_select')!.value).toEqual('2');
+      expect(node.querySelector('#root_content')).not.toBeInTheDocument();
+
+      // Switch back to "Advanced Configuration" (index 1)
+      await user.selectOptions(node.querySelector<HTMLSelectElement>('#root__oneof_select')!, '1');
+
+      // The content field should be restored with defaults
+      expect(node.querySelector<HTMLSelectElement>('#root__oneof_select')!.value).toEqual('1');
+      contentInput = node.querySelector<HTMLInputElement>('#root_content');
+      expect(contentInput).toBeInTheDocument();
+      // BUG: Without the fix, this would be empty string or undefined
+      expect(contentInput!.value).toEqual('placeholder');
+
+      // Also verify the final formData has correct values
+      const lastFormData = onChangeCalls[onChangeCalls.length - 1].event.formData;
+      expect(lastFormData.types).toEqual('advanced');
+      expect(lastFormData.content).toEqual('placeholder');
+    });
+    it('should allow switching to null option in oneOf', async () => {
+      // This test verifies that switching to a null option in oneOf works correctly.
+      // Without the fix, the form would revert back to the previous option.
+      // NOTE: This bug only manifests in controlled forms where parent updates formData prop.
+      const schema: RJSFSchema = {
+        type: 'object',
+        title: 'Configuration',
+        oneOf: [
+          {
+            title: 'Default Configuration',
+            type: 'object',
+            properties: {
+              types: { const: 'default', title: 'Types' },
+              content: { type: 'string', title: 'Content' },
+            },
+            required: ['types'],
+          },
+          {
+            title: 'Advanced Configuration',
+            type: 'object',
+            properties: {
+              types: { const: 'advanced', title: 'Types' },
+              content: { type: 'string', title: 'Content' },
+            },
+            required: ['types'],
+          },
+          { title: 'No Configuration', type: 'null' },
+        ],
+        default: { types: 'advanced', content: 'placeholder' },
+      };
+
+      const onChangeCalls: { event: IChangeEvent; id?: string }[] = [];
+      let currentFormData: unknown = undefined;
+
+      const { node, rerender } = createFormComponent({
+        ref: createRef(),
+        schema,
+        onChange: (event: IChangeEvent, id?: string) => {
+          onChangeCalls.push({ event, id });
+          currentFormData = event.formData;
+        },
+        experimental_defaultFormStateBehavior: { emptyObjectFields: 'populateAllDefaults' },
+      });
+
+      // Should start with "Advanced Configuration" (index 1)
+      expect(node.querySelector<HTMLSelectElement>('#root__oneof_select')!.value).toEqual('1');
+      expect(node.querySelector('#root_content')).toBeInTheDocument();
+
+      // Switch to "No Configuration" (null option, index 2)
+      await user.selectOptions(node.querySelector<HTMLSelectElement>('#root__oneof_select')!, '2');
+
+      // Simulate controlled form behavior by re-rendering with new formData
+      rerender({
+        ref: createRef(),
+        schema,
+        formData: currentFormData,
+        experimental_defaultFormStateBehavior: { emptyObjectFields: 'populateAllDefaults' },
+      });
+
+      // BUG: Without the fix, the form would revert back to index 1
+      expect(node.querySelector<HTMLSelectElement>('#root__oneof_select')!.value).toEqual('2');
+      // Content field should not exist for null option
+      expect(node.querySelector('#root_content')).not.toBeInTheDocument();
+
+      // Verify formData is null or undefined (both valid for null option)
+      const lastFormData = onChangeCalls[onChangeCalls.length - 1].event.formData;
+      expect(lastFormData == null).toBe(true);
+    });
+    it('Should modify anyOf definition references when the defaults are set.', async () => {
+      const schema: RJSFSchema = {
+        definitions: {
+          option1: {
+            properties: {
+              first: {
+                type: 'string',
+              },
+            },
+          },
+          option2: {
+            properties: {
+              second: {
+                type: 'string',
+              },
+            },
+          },
+        },
+        properties: {
+          any_of_field: {
+            anyOf: [
+              {
+                $ref: '#/definitions/option1',
+              },
+              {
+                $ref: '#/definitions/option2',
+              },
+            ],
+            default: {
+              second: 'second!',
+            },
+          },
+        },
+        type: 'object',
+      };
+
+      const { node, onChange } = createFormComponent({
+        schema,
+      });
+
+      const secondInputID = '#root_any_of_field_second';
+      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'second!');
+
+      await user.clear(node.querySelector(secondInputID)!);
+      await user.type(node.querySelector(secondInputID)!, 'changed!');
+
+      expectToHaveBeenCalledWithFormData(
+        onChange,
+        {
+          any_of_field: {
+            second: 'changed!',
+          },
+        },
+        'root_any_of_field_second',
+      );
+
+      expect(node.querySelector(secondInputID)).toHaveAttribute('value', 'changed!');
+    });
+    it('Should modify oneOf object with references when the defaults are set.', async () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        $defs: {
+          protocol: {
+            type: 'string',
+            enum: ['fast', 'balanced', 'stringent'],
+            default: 'fast',
+          },
+        },
+        oneOf: [
+          {
+            properties: {
+              protocol: {
+                $ref: '#/$defs/protocol',
+              },
+            },
+          },
+          {
+            properties: {
+              something: {
+                type: 'number',
+              },
+            },
+          },
+        ],
+      };
+
+      const { node, onChange } = createFormComponent({
+        schema,
+      });
+
+      const protocolInputID = '#root_protocol';
+      expect(node.querySelector(protocolInputID)).toHaveValue('0');
+
+      await user.selectOptions(node.querySelector(protocolInputID)!, '1');
+
+      expectToHaveBeenCalledWithFormData(
+        onChange,
+        {
+          protocol: 'balanced',
+        },
+        'root_protocol',
+      );
+
+      expect(node.querySelector(protocolInputID)).toHaveValue('1');
+    });
+    describe('Should modify oneOf radio button when the defaults are set.', () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        properties: {
+          a: {
+            type: ['boolean', 'null'],
+            default: null,
+            oneOf: [
+              {
+                const: false,
+                title: 'No',
+              },
+              {
+                const: null,
+                title: 'N/A',
+              },
+            ],
+          },
+        },
+        allOf: [
+          {
+            if: {
+              required: ['a'],
+              properties: {
+                a: {
+                  const: false,
+                },
+              },
+            },
+            then: {
+              required: ['b'],
+              properties: {
+                b: {
+                  type: 'string',
+                },
+              },
+            },
+          },
+        ],
+      };
+
+      const uiSchema: UiSchema = {
+        a: {
+          'ui:widget': 'radio',
+          'ui:label': false,
+        },
+      };
+      const notApplicableInputID = '#root_a-1';
+      const NoInputID = '#root_a-0';
+
+      it('Test with default constAsDefaults', async () => {
+        const { node, onChange } = createFormComponent({
+          schema,
+          uiSchema,
+        });
+
+        expect(node.querySelector(notApplicableInputID)).toBeChecked();
+
+        await user.click(node.querySelector(NoInputID)!);
+
+        expectToHaveBeenCalledWithFormData(onChange, { a: false }, 'root_a');
+
+        expect(node.querySelector(NoInputID)).toBeChecked();
+        expect(node.querySelector(notApplicableInputID)).not.toBeChecked();
+        expect(node.querySelector('#root_b')).toBeInTheDocument();
+      });
+      it('Test with constAsDefaults set to "never"', async () => {
+        const { node, onChange } = createFormComponent({
+          schema,
+          uiSchema,
+          experimental_defaultFormStateBehavior: {
+            constAsDefaults: 'never',
+          },
+        });
+
+        expect(node.querySelector(notApplicableInputID)).toBeChecked();
+
+        await user.click(node.querySelector(NoInputID)!);
+
+        expectToHaveBeenCalledWithFormData(onChange, { a: false }, 'root_a');
+
+        expect(node.querySelector(NoInputID)).toBeChecked();
+        expect(node.querySelector(notApplicableInputID)).not.toBeChecked();
+        expect(node.querySelector('#root_b')).toBeInTheDocument();
+      });
+    });
+  });
+
+  describe('Blur handler', () => {
+    it('should call provided blur handler on form input blur event', async () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        properties: {
+          foo: {
+            type: 'string',
+          },
+        },
+      };
+      const formData = {
+        foo: '',
+      };
+      const onBlur = vi.fn();
+      const { node } = createFormComponent({ schema, formData, onBlur });
+
+      const input = node.querySelector('[type=text]')!;
+      await user.type(input, 'new');
+      await user.tab();
+
+      expect(onBlur).toHaveBeenLastCalledWith(input.id, 'new');
+    });
+  });
+
+  describe('Focus handler', () => {
+    it('should call provided focus handler on form input focus event', async () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        properties: {
+          foo: {
+            type: 'string',
+          },
+        },
+      };
+      const formData = {
+        foo: 'new',
+      };
+      const onFocus = vi.fn();
+      const { node } = createFormComponent({ schema, formData, onFocus });
+
+      const input = node.querySelector('[type=text]')!;
+      await user.click(input);
+
+      expect(onFocus).toHaveBeenLastCalledWith(input.id, 'new');
+    });
+  });
+
+  describe('Error handler', () => {
+    it('should call provided error handler on validation errors', async () => {
+      const schema: RJSFSchema = {
+        type: 'object',
+        properties: {
+          foo: {
+            type: 'string',
+            minLength: 1,
+          },
+        },
+      };
+      const formData = {
+        foo: '',
+      };
+      const { node, onError } = createFormComponent({ schema, formData });
+
+      await submitForm(node, user);
+
+      expect(onError).toHaveBeenCalledTimes(1);
+    });
+  });
+});
