+++ packages/core/test/ObjectField.test.tsx
@@ -15,8 +15,8 @@ import userEvent from '@testing-library/user-event';
 import type Form from '../src';
 import ObjectField from '../src/components/fields/ObjectField';
 import SchemaField from '../src/components/fields/SchemaField';
-import { TextWidgetTest } from './StringField.test';
 import { createFormComponent, expectToHaveBeenCalledWithFormData, submitForm } from './testUtils';
+import { TextWidgetTest } from './TextWidgetTest';
 
 const user = userEvent.setup();
